Revert "Create deck sheet (#55)"
This reverts commit 0b0ad3043d8141f7e2fa230a56eb7472ec3f0f58.
