+++ lang/en.json
@@ -26,12 +26,9 @@
         "Skills": "Skills",
         "NPC": "NPC Details",
         "Ship": "Ship Details",
-        "Crew": "Crew",
-        "Cargo": "Cargo",
         "Drone": "Drone Details",
         "Mech": "Mech Details",
-        "Vehicle": "Vehicle Details",
-        "Programs": "Program Details"
+        "Vehicle": "Vehicle Details"
       },
       "base": {
         "FIELDS": {
@@ -176,39 +173,6 @@
           "range": { "label": "Range" },
           "model": { "label": "Model" },
           "moveType": { "label": "Move Type" }
-        }
-      },
-      "Cyberdeck": {
-        "FIELDS": {
-          "health": {
-            "value": { "label": "Shielding"},
-            "label": "Shielding" 
-          },
-          "access": {
-            "value": { "label": "Access"},
-            "label": "Access"
-          },
-          "hackerHP": {
-            "value": { "label": "HP"},
-            "label": "HP" 
-          },
-          "memory": {
-            "value": { "label": "Memory"},
-            "label": "Memory"
-          },
-          "cpu": {
-            "value": { "label": "CPU"},
-            "label": "CPU"
-          },
-          "bonusAccess": { "label" : "Bonus Access"},
-          "baseShielding": {"label" : "Base Shielding"},
-          "bonusShielding": { "label" : "Bonus Shielding"},
-          "neuralBuffer": {"label" : "Neural Buffer"},
-          "wirelessConnectionPenalty": {"label": "Wireless Connection Penalty (-2)"},
-          "crownPenalty": { "label": "Crown Penalty (-1)"},
-          "cost": {"label": "Cost"},
-          "encumberance": {"label": "Encumberance"},
-          "hacker": {"label": "Hacker"}
         }
       }
     },
@@ -830,7 +794,6 @@
       "rollables": "Rollables",
       "roll-skill": "Roll Skill",
       "hit-dice": "Hit Dice",
-      "carriedGear": "Carried Gear",
       "hd": "Hit Dice per level (without CON mod)",
       "command-points": "Command Points",
       "npc-command-points": "NPC Ship CP",
@@ -1068,13 +1031,10 @@
           "role": "Station",
           "crew": "Crew"
         },
-        "cargo-name": "Cargo",
-        "cargo-amount": "Amount",
-        "cargo-cost": "Cost/Unit"
+        "cargo-name": "Cargo Name / Description"
       },
       "cyberdeck": {
         "shielding": "Shielding",
-        "health": "HP",
         "hacker": "Hacker",
         "bonusAccess": "Bonus Access",
         "memory": "Memory",
@@ -1084,10 +1044,7 @@
         "neuralBuffer": "Neural Buffer",
         "wirelessConnectionPenalty": "Wireless(-2)",
         "crownPenalty": "Crown (-1)",
-        "run": "Run/ Activate Program",
-        "refreshShielding": "Refresh Shielding",
-        "refreshAccess": "Refresh Access",
-        "hackerHP": "Hacker HP"
+        "run": "Run/ Activate Program"
       },
       "program": {
         "type": "Type",
@@ -1107,8 +1064,7 @@
         "targetAllowedShort": "Targets",
         "subjectType": "Subject Type",
         "attackRoll": "Attack Roll",
-        "skillRoll": "Skill Roll",
-        "run": "Run Program"
+        "skillRoll": "Skill Roll"
       },
       "tweaks": "Tweaks"
     },
@@ -1196,8 +1152,7 @@
       "locationReadied": "Readied",
       "range": "Range",
       "favorite": "Favorite",
-      "noEncReadied": "No Encumbrance Readied",
-      "type": "Type"
+      "noEncReadied": "No Encumbrance Readied"
     },
     "weapon": {
       "shock": {
