+++ module/data/actors/base-vehicle.mjs
@@ -54,7 +54,7 @@ export default class SWNRResource  extends foundry.abstract.TypeDataModel {
     schema.armor = SWNShared.resourceField(1,1);
     schema.speed = SWNShared.requiredNumber(1);
     schema.crew = SWNShared.rangeResourceField(1,1,1);
-    schema.crewMembers = new fields.ArrayField(new fields.DocumentIdField());
+    schema.crewMembers = new fields.ArrayField(new fields.DocumentUUIDField());
     schema.tl = SWNShared.requiredNumber(5);
 
     schema.description = new fields.HTMLField();
@@ -64,20 +64,5 @@ export default class SWNRResource  extends foundry.abstract.TypeDataModel {
     schema.mass = SWNShared.resourceField(1,1,true);
     schema.hardpoints = SWNShared.resourceField(1,1,true);
     return schema;
-  }
-
-  prepareDerivedData() {
-    this.health.percentage = Math.clamp((this.health.value * 100) / this.health.max, 0, 100);
-    this.armor.percentage = Math.clamp((this.armor.value * 100) / this.armor.max, 0, 100);
-
-    this.carriedGear = this.parent.items.filter((item) => 
-      item.type === "item" || item.type === "weapon" || item.type === "armor");
-    this.carriedGear.sort((a, b) => {
-      if (a.type == b.type) {
-        return a.name.localeCompare(b.name);
-      } else {
-        return a.type.localeCompare(b.type);
-      }
-    });
   }
 }
