+++ module/sheets/actor-sheet.mjs
@@ -2,8 +2,6 @@ import { prepareActiveEffectCategories } from '../helpers/effects.mjs';
 import { getGameSettings } from '../helpers/register-settings.mjs';
 import { headerFieldWidget, groupFieldWidget } from '../helpers/handlebar.mjs';
 import { initSkills, initCompendSkills, calcMod } from '../helpers/utils.mjs';
-import { SWNBaseSheet } from './base-sheet.mjs';
-
 
 const { api, sheets } = foundry.applications;
 
@@ -11,9 +9,12 @@ const { api, sheets } = foundry.applications;
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheetV2}
  */
-export class SWNActorSheet extends SWNBaseSheet {
+export class SWNActorSheet extends api.HandlebarsApplicationMixin(
+  sheets.ActorSheetV2
+) {
   constructor(options = {}) {
     super(options);
+    this.#dragDrop = this.#createDragDropHandlers();
   }
 
   // Private properties
@@ -89,7 +90,7 @@ export class SWNActorSheet extends SWNBaseSheet {
       template: 'systems/swnr/templates/actor/skills.hbs',
     },
     // FRAGMENTS
-    itemsList: {
+    itemsList: { 
       template: 'systems/swnr/templates/actor/fragments/items-list.hbs',
     },
     skillFrag: {
@@ -111,7 +112,7 @@ export class SWNActorSheet extends SWNBaseSheet {
     super._configureRenderOptions(options);
 
     //wsai adding to allow setting default tab
-    options.defaultTab = 'gear';
+    options.defaultTab= 'gear';
 
     // Not all parts always render
     options.parts = ['header', 'tabs', 'biography'];
@@ -121,11 +122,11 @@ export class SWNActorSheet extends SWNBaseSheet {
     switch (this.document.type) {
       case 'character':
         // ws AI removing skills for now: ,'skills'
-        options.parts.push('combat', 'features', 'gear', 'powers', 'effects');
+        options.parts.push('combat','features', 'gear', 'powers', 'effects');
         options.defaultTab = 'combat';
         break;
       case 'npc':
-        options.parts.push('npc', 'gear', 'effects');
+        options.parts.push('npc','gear', 'effects');
         options.defaultTab = 'npc';
         break;
     }
@@ -295,7 +296,7 @@ export class SWNActorSheet extends SWNBaseSheet {
       // Append to gear.
       if (i.type === 'item') {
         items.push(i);
-      }
+      } 
       else if (i.type === 'armor') {
         items.push(i); //TODO fix to new array
       }
@@ -306,7 +307,7 @@ export class SWNActorSheet extends SWNBaseSheet {
       else if (i.type === 'feature') {
         features.push(i);
       }
-      else if (i.type === 'cyberware') {
+      else if (i.type === 'cyberware') {  
         cyberware.push(i);
       }
       // Append to powers.
@@ -329,7 +330,7 @@ export class SWNActorSheet extends SWNBaseSheet {
     context.cyberware = cyberware.sort((a, b) => {
       if (a.system.type === "none" && b.system.type !== "none") return -1; // a comes first
       if (b.system.type === "none" && a.system.type !== "none") return 1;  // b comes first
-
+      
       if ((a.system.type || 0) < (b.system.type || 0)) { return -1; }
       if ((a.system.type || 0) > (b.system.type || 0)) { return 1; }
       return 0;
@@ -345,14 +346,15 @@ export class SWNActorSheet extends SWNBaseSheet {
    * @override
    */
   _onRender(context, options) {
-    super._onRender(context, options);
+    this.#dragDrop.forEach((d) => d.bind(this.element));
+    this.#disableOverrides();
 
     // You may want to add other special handling here
     // Foundry comes with a large number of utility classes, e.g. SearchFilter
     // That you may want to implement yourself.
 
     // NOTE that 'click' events should just be actions like roll, rest, etc.
-    this.element.querySelectorAll(".location-selector").forEach((d) =>
+    this.element.querySelectorAll(".location-selector").forEach((d) => 
       d.addEventListener('change', this._onLocationChange.bind(this)));
 
     // Toggle lock related elements after render depending on the lock state
@@ -385,6 +387,106 @@ export class SWNActorSheet extends SWNBaseSheet {
    **************/
 
   /**
+   * Handle changing a Document's image.
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @returns {Promise}
+   * @protected
+   */
+  static async _onEditImage(event, target) {
+    const attr = target.dataset.edit;
+    const current = foundry.utils.getProperty(this.document, attr);
+    const { img } =
+      this.document.constructor.getDefaultArtwork?.(this.document.toObject()) ??
+      {};
+    const fp = new FilePicker({
+      current,
+      type: 'image',
+      redirectToRoot: img ? [img] : [],
+      callback: (path) => {
+        this.document.update({ [attr]: path });
+      },
+      top: this.position.top + 40,
+      left: this.position.left + 10,
+    });
+    return fp.browse();
+  }
+
+  /**
+   * Renders an embedded document's sheet
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _viewDoc(event, target) {
+    const doc = this._getEmbeddedDocument(target);
+    doc.sheet.render(true);
+  }
+
+  /**
+   * Handles item deletion
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _deleteDoc(event, target) {
+    const doc = this._getEmbeddedDocument(target);
+    await doc.delete();
+  }
+
+  /**
+   * Handle creating a new Owned Item or ActiveEffect for the actor using initial data defined in the HTML dataset
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @private
+   */
+  static async _createDoc(event, target) {
+    // Retrieve the configured document class for Item or ActiveEffect
+    const docCls = getDocumentClass(target.dataset.documentClass);
+    // Prepare the document creation data by initializing it a default name.
+    const docData = {
+      name: docCls.defaultName({
+        // defaultName handles an undefined type gracefully
+        type: target.dataset.type,
+        parent: this.actor,
+      }),
+    };
+    // Loop through the dataset and add it to our docData
+    for (const [dataKey, value] of Object.entries(target.dataset)) {
+      // These data attributes are reserved for the action handling
+      if (['action', 'documentClass'].includes(dataKey)) continue;
+      // Nested properties require dot notation in the HTML, e.g. anything with `system`
+      // An example exists in powers.hbs, with `data-system.power-level`
+      // which turns into the dataKey 'system.powerLevel'
+      foundry.utils.setProperty(docData, dataKey, value);
+    }
+
+    // Finally, create the embedded document!
+    await docCls.create(docData, { parent: this.actor });
+  }
+
+  /**
+   * Determines effect parent to pass to helper
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @private
+   */
+  static async _toggleEffect(event, target) {
+    const effect = this._getEmbeddedDocument(target);
+    await effect.update({ disabled: !effect.disabled });
+  }
+
+  /**
    * Handle clickable rolls.
    *
    * @this SWNActorSheet
@@ -401,21 +503,21 @@ export class SWNActorSheet extends SWNBaseSheet {
         rejectClose: false,
         modal: true,
         buttons: [
-          {
-            icon: 'fas fa-check',
-            label: "Yes",
-            action: "normal",
-          },
-          {
-            icon: 'fas fa-check',
-            label: "Yes, but no HP",
-            action: "no_hp",
-          },
-          {
-            icon: 'fas fa-times',
-            label: "No",
-            action: "no",
-          },
+        {
+          icon: 'fas fa-check',
+          label: "Yes",
+          action: "normal",
+        },
+        {
+          icon: 'fas fa-check',
+          label: "Yes, but no HP",
+          action: "no_hp",
+        },
+        {
+          icon: 'fas fa-times',
+          label: "No",        
+          action: "no",
+        },
         ]
       })
       if (rest === "no") {
@@ -455,10 +557,10 @@ export class SWNActorSheet extends SWNBaseSheet {
 
   static async _onScene(event, _target) {
     event.preventDefault();
-    let update = { system: { effort: { scene: 0 } } };
+    let update = { system: { effort: { scene: 0 }}};
     if (this.actor.type === "character") {
       update["tweak.extraEffort.scene"] = 0;
-    }
+    } 
     await this.actor.update(update);
     this._resetSoak();
   }
@@ -476,7 +578,7 @@ export class SWNActorSheet extends SWNBaseSheet {
     }
 
   }
-
+  
   async _resetSoak() {
     if (game.settings.get("swnr", "useCWNArmor")) {
       if (this.actor.type == "npc") {
@@ -501,7 +603,7 @@ export class SWNActorSheet extends SWNBaseSheet {
       }
     }
   }
-
+  
   /**
    * Skill roll
    */
@@ -514,7 +616,7 @@ export class SWNActorSheet extends SWNBaseSheet {
 
   static async _onSkillUp(event, target) {
     event.preventDefault();
-
+  
     const skillID = target.dataset.itemId;
     const skill = this.actor.items.get(skillID);
     const rank = skill.system.rank;
@@ -543,12 +645,12 @@ export class SWNActorSheet extends SWNBaseSheet {
     const skillCost = rank + 2;
     const isPsy =
       skill.system.source.toLocaleLowerCase() ===
-        game.i18n.localize("swnr.skills.labels.psionic").toLocaleLowerCase()
+      game.i18n.localize("swnr.skills.labels.psionic").toLocaleLowerCase()
         ? true
         : false;
     const skillPointsAvail = isPsy
       ? this.actor.system.unspentPsySkillPoints +
-      this.actor.system.unspentSkillPoints
+        this.actor.system.unspentSkillPoints
       : this.actor.system.unspentSkillPoints;
     if (skillCost > skillPointsAvail) {
       ui.notifications?.error(
@@ -591,7 +693,7 @@ export class SWNActorSheet extends SWNBaseSheet {
     const armorID = target.dataset.itemId;
     const armor = this.actor.items.get(armorID);
     const use = armor.system.use;
-    await armor.update({ system: { use: !use } });
+    await armor.update({ system: { use: !use }});
   }
 
    /**
@@ -638,23 +740,22 @@ export class SWNActorSheet extends SWNBaseSheet {
 
     const _resp = await foundry.applications.api.DialogV2.prompt(
       {
-        window: {
+        window: { 
           title: game.i18n.format("swnr.dialog.add-bulk-skills",
             { actor: this.actor.name }
-          )
-        },
+          )   }, 
         content,
         modal: true,
         rejectClose: false,
-        ok:
-        {
-          action: "addSkills",
-          label: game.i18n.localize("swnr.dialog.add-skills"),
-          callback: _addSkills,
-        },
+        ok: 
+          {
+            action: "addSkills",
+            label: game.i18n.localize("swnr.dialog.add-skills"),
+            callback: _addSkills,
+          },
       }
     );
-  }
+  }    
 
   /**
    * Handle clickable rolls.
@@ -691,7 +792,7 @@ export class SWNActorSheet extends SWNBaseSheet {
   static async _onRollStats(event, _target) {
     event.preventDefault();
     const dice = await foundry.applications.api.DialogV2.wait({
-      window: { title: game.i18n.localize("swnr.chat.statRoll", { name: this.actor.name }) },
+      window: { title: game.i18n.localize("swnr.chat.statRoll", {name: this.actor.name}) },
       content: game.i18n.localize("swnr.dialog.statMethod"),
       modal: true,
       buttons: [
@@ -713,7 +814,7 @@ export class SWNActorSheet extends SWNBaseSheet {
     const formula = new Array(6).fill(dice).join("+");
     const roll = new Roll(formula);
     await roll.roll({ async: true });
-    const stats = {};
+    const stats= {};
     ["str", "dex", "con", "int", "wis", "cha"].map((k, i) => {
       stats[k] = {
         dice: roll.dice[i].results,
@@ -837,5 +938,353 @@ export class SWNActorSheet extends SWNBaseSheet {
         callback: _doAdd,
       }
     });
+  }
+
+  /** Helper Functions */
+
+  /**
+   * Fetches the embedded document representing the containing HTML element
+   *
+   * @param {HTMLElement} target    The element subject to search
+   * @returns {Item | ActiveEffect} The embedded Item or ActiveEffect
+   */
+  _getEmbeddedDocument(target) {
+    const docRow = target.closest('li[data-document-class]');
+    if (docRow.dataset.documentClass === 'Item') {
+      return this.actor.items.get(docRow.dataset.itemId);
+    } else if (docRow.dataset.documentClass === 'ActiveEffect') {
+      const parent =
+        docRow.dataset.parentId === this.actor.id
+          ? this.actor
+          : this.actor.items.get(docRow?.dataset.parentId);
+      return parent.effects.get(docRow?.dataset.effectId);
+    } else return console.warn('Could not find document class');
+  }
+
+  /***************
+   *
+   * Drag and Drop
+   *
+   ***************/
+
+  /**
+   * Define whether a user is able to begin a dragstart workflow for a given drag selector
+   * @param {string} selector       The candidate HTML selector for dragging
+   * @returns {boolean}             Can the current user drag this selector?
+   * @protected
+   */
+  _canDragStart(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
+
+  /**
+   * Define whether a user is able to conclude a drag-and-drop workflow for a given drop selector
+   * @param {string} selector       The candidate HTML selector for the drop target
+   * @returns {boolean}             Can the current user drop on this selector?
+   * @protected
+   */
+  _canDragDrop(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
+
+  /**
+   * Callback actions which occur at the beginning of a drag start workflow.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  _onDragStart(event) {
+    const docRow = event.currentTarget.closest('li');
+    if ('link' in event.target.dataset) return;
+
+    // Chained operation
+    let dragData = this._getEmbeddedDocument(docRow)?.toDragData();
+
+    if (!dragData) return;
+
+    // Set data transfer
+    event.dataTransfer.setData('text/plain', JSON.stringify(dragData));
+  }
+
+  /**
+   * Callback actions which occur when a dragged element is over a drop target.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  _onDragOver(event) {}
+
+  /**
+   * Callback actions which occur when a dragged element is dropped on a target.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  async _onDrop(event) {
+    const data = TextEditor.getDragEventData(event);
+    const actor = this.actor;
+    const allowed = Hooks.call('dropActorSheetData', actor, this, data);
+    if (allowed === false) return;
+
+    // Handle different data types
+    switch (data.type) {
+      case 'ActiveEffect':
+        return this._onDropActiveEffect(event, data);
+      case 'Actor':
+        return this._onDropActor(event, data);
+      case 'Item':
+        return this._onDropItem(event, data);
+      case 'Folder':
+        return this._onDropFolder(event, data);
+    }
+  }
+
+  /**
+   * Handle the dropping of ActiveEffect data onto an Actor Sheet
+   * @param {DragEvent} event                  The concluding DragEvent which contains drop data
+   * @param {object} data                      The data transfer extracted from the event
+   * @returns {Promise<ActiveEffect|boolean>}  The created ActiveEffect object or false if it couldn't be created.
+   * @protected
+   */
+  async _onDropActiveEffect(event, data) {
+    const aeCls = getDocumentClass('ActiveEffect');
+    const effect = await aeCls.fromDropData(data);
+    if (!this.actor.isOwner || !effect) return false;
+    if (effect.target === this.actor)
+      return this._onSortActiveEffect(event, effect);
+    return aeCls.create(effect, { parent: this.actor });
+  }
+
+  /**
+   * Handle a drop event for an existing embedded Active Effect to sort that Active Effect relative to its siblings
+   *
+   * @param {DragEvent} event
+   * @param {ActiveEffect} effect
+   */
+  async _onSortActiveEffect(event, effect) {
+    /** @type {HTMLElement} */
+    const dropTarget = event.target.closest('[data-effect-id]');
+    if (!dropTarget) return;
+    const target = this._getEmbeddedDocument(dropTarget);
+
+    // Don't sort on yourself
+    if (effect.uuid === target.uuid) return;
+
+    // Identify sibling items based on adjacent HTML elements
+    const siblings = [];
+    for (const el of dropTarget.parentElement.children) {
+      const siblingId = el.dataset.effectId;
+      const parentId = el.dataset.parentId;
+      if (
+        siblingId &&
+        parentId &&
+        (siblingId !== effect.id || parentId !== effect.parent.id)
+      )
+        siblings.push(this._getEmbeddedDocument(el));
+    }
+
+    // Perform the sort
+    const sortUpdates = SortingHelpers.performIntegerSort(effect, {
+      target,
+      siblings,
+    });
+
+    // Split the updates up by parent document
+    const directUpdates = [];
+
+    const grandchildUpdateData = sortUpdates.reduce((items, u) => {
+      const parentId = u.target.parent.id;
+      const update = { _id: u.target.id, ...u.update };
+      if (parentId === this.actor.id) {
+        directUpdates.push(update);
+        return items;
+      }
+      if (items[parentId]) items[parentId].push(update);
+      else items[parentId] = [update];
+      return items;
+    }, {});
+
+    // Effects-on-items updates
+    for (const [itemId, updates] of Object.entries(grandchildUpdateData)) {
+      await this.actor.items
+        .get(itemId)
+        .updateEmbeddedDocuments('ActiveEffect', updates);
+    }
+
+    // Update on the main actor
+    return this.actor.updateEmbeddedDocuments('ActiveEffect', directUpdates);
+  }
+
+  /**
+   * Handle dropping of an Actor data onto another Actor sheet
+   * @param {DragEvent} event            The concluding DragEvent which contains drop data
+   * @param {object} data                The data transfer extracted from the event
+   * @returns {Promise<object|boolean>}  A data object which describes the result of the drop, or false if the drop was
+   *                                     not permitted.
+   * @protected
+   */
+  async _onDropActor(event, data) {
+    if (!this.actor.isOwner) return false;
+  }
+
+  /* -------------------------------------------- */
+
+  /**
+   * Handle dropping of an item reference or item data onto an Actor Sheet
+   * @param {DragEvent} event            The concluding DragEvent which contains drop data
+   * @param {object} data                The data transfer extracted from the event
+   * @returns {Promise<Item[]|boolean>}  The created or updated Item instances, or false if the drop was not permitted.
+   * @protected
+   */
+  async _onDropItem(event, data) {
+    if (!this.actor.isOwner) return false;
+    const item = await Item.implementation.fromDropData(data);
+
+    // Handle item sorting within the same Actor
+    if (this.actor.uuid === item.parent?.uuid)
+      return this._onSortItem(event, item);
+
+    // Create the owned item
+    return this._onDropItemCreate(item, event);
+  }
+
+  /**
+   * Handle dropping of a Folder on an Actor Sheet.
+   * The core sheet currently supports dropping a Folder of Items to create all items as owned items.
+   * @param {DragEvent} event     The concluding DragEvent which contains drop data
+   * @param {object} data         The data transfer extracted from the event
+   * @returns {Promise<Item[]>}
+   * @protected
+   */
+  async _onDropFolder(event, data) {
+    if (!this.actor.isOwner) return [];
+    const folder = await Folder.implementation.fromDropData(data);
+    if (folder.type !== 'Item') return [];
+    const droppedItemData = await Promise.all(
+      folder.contents.map(async (item) => {
+        if (!(document instanceof Item)) item = await fromUuid(item.uuid);
+        return item;
+      })
+    );
+    return this._onDropItemCreate(droppedItemData, event);
+  }
+
+  /**
+   * Handle the final creation of dropped Item data on the Actor.
+   * This method is factored out to allow downstream classes the opportunity to override item creation behavior.
+   * @param {object[]|object} itemData      The item data requested for creation
+   * @param {DragEvent} event               The concluding DragEvent which provided the drop data
+   * @returns {Promise<Item[]>}
+   * @private
+   */
+  async _onDropItemCreate(itemData, event) {
+    itemData = itemData instanceof Array ? itemData : [itemData];
+    return this.actor.createEmbeddedDocuments('Item', itemData);
+  }
+
+  /**
+   * Handle a drop event for an existing embedded Item to sort that Item relative to its siblings
+   * @param {Event} event
+   * @param {Item} item
+   * @private
+   */
+  _onSortItem(event, item) {
+    // Get the drag source and drop target
+    const items = this.actor.items;
+    const dropTarget = event.target.closest('[data-item-id]');
+    if (!dropTarget) return;
+    const target = items.get(dropTarget.dataset.itemId);
+
+    // Don't sort on yourself
+    if (item.id === target.id) return;
+
+    // Identify sibling items based on adjacent HTML elements
+    const siblings = [];
+    for (let el of dropTarget.parentElement.children) {
+      const siblingId = el.dataset.itemId;
+      if (siblingId && siblingId !== item.id)
+        siblings.push(items.get(el.dataset.itemId));
+    }
+
+    // Perform the sort
+    const sortUpdates = SortingHelpers.performIntegerSort(item, {
+      target,
+      siblings,
+    });
+    const updateData = sortUpdates.map((u) => {
+      const update = u.update;
+      update._id = u.target._id;
+      return update;
+    });
+
+    // Perform the update
+    return this.actor.updateEmbeddedDocuments('Item', updateData);
+  }
+
+  /** The following pieces set up drag handling and are unlikely to need modification  */
+
+  /**
+   * Returns an array of DragDrop instances
+   * @type {DragDrop[]}
+   */
+  get dragDrop() {
+    return this.#dragDrop;
+  }
+
+  // This is marked as private because there's no real need
+  // for subclasses or external hooks to mess with it directly
+  #dragDrop;
+
+  /**
+   * Create drag-and-drop workflow handlers for this Application
+   * @returns {DragDrop[]}     An array of DragDrop handlers
+   * @private
+   */
+  #createDragDropHandlers() {
+    return this.options.dragDrop.map((d) => {
+      d.permissions = {
+        dragstart: this._canDragStart.bind(this),
+        drop: this._canDragDrop.bind(this),
+      };
+      d.callbacks = {
+        dragstart: this._onDragStart.bind(this),
+        dragover: this._onDragOver.bind(this),
+        drop: this._onDrop.bind(this),
+      };
+      return new DragDrop(d);
+    });
+  }
+
+  /********************
+   *
+   * Actor Override Handling
+   *
+   ********************/
+
+  /**
+   * Submit a document update based on the processed form data.
+   * @param {SubmitEvent} event                   The originating form submission event
+   * @param {HTMLFormElement} form                The form element that was submitted
+   * @param {object} submitData                   Processed and validated form data to be used for a document update
+   * @returns {Promise<void>}
+   * @protected
+   * @override
+   */
+  async _processSubmitData(event, form, submitData) {
+    const overrides = foundry.utils.flattenObject(this.actor.overrides);
+    for (let k of Object.keys(overrides)) delete submitData[k];
+    await this.document.update(submitData);
+  }
+
+  /**
+   * Disables inputs subject to active effects
+   */
+  #disableOverrides() {
+    const flatOverrides = foundry.utils.flattenObject(this.actor.overrides);
+    for (const override of Object.keys(flatOverrides)) {
+      const input = this.element.querySelector(`[name="${override}"]`);
+      if (input) {
+        input.disabled = true;
+      }
+    }
   }
 }
