+++ module/sheets/cyberdeck-sheet.mjs
@@ -1,7 +1,4 @@
-import { getGameSettings } from '../helpers/register-settings.mjs';
-import { headerFieldWidget, groupFieldWidget } from '../helpers/handlebar.mjs';
-
-import { SWNBaseSheet } from './base-sheet.mjs';
+import { prepareActiveEffectCategories } from '../helpers/effects.mjs';
 
 const { api, sheets } = foundry.applications;
 
@@ -9,17 +6,20 @@ const { api, sheets } = foundry.applications;
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheetV2}
  */
-export class SWNCyberdeckSheet extends SWNBaseSheet {
+export class SWNCyberdeckSheet extends api.HandlebarsApplicationMixin(
+  sheets.ActorSheetV2
+) {
   constructor(options = {}) {
     super(options);
+    this.#dragDrop = this.#createDragDropHandlers();
   }
 
   /** @override */
   static DEFAULT_OPTIONS = {
     classes: ['swnr', 'actor'],
     position: {
-      width: 800,
-      height: 1000,
+      width: 600,
+      height: 600,
     },
     actions: {
       onEditImage: this._onEditImage,
@@ -28,11 +28,6 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
       deleteDoc: this._deleteDoc,
       toggleEffect: this._toggleEffect,
       roll: this._onRoll,
-      hackerDelete: this._onHackerDelete,
-      hackerRoll: this._onHackerRoll,
-      runProgram: this._onActivateProgram,
-      refreshAccess: this._onRefreshAccess,
-      refreshShielding: this._onRefreshShielding
     },
     // Custom property that's merged into `this.options`
     dragDrop: [{ dragSelector: '[data-drag]', dropSelector: null }],
@@ -44,39 +39,45 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
   /** @override */
   static PARTS = {
     header: {
-      template: 'systems/swnr/templates/actor/cyberdeck/cyberdeckHeader.hbs',
+      template: 'systems/swnr/templates/actor/vehicle/header.hbs',
     },
     tabs: {
       // Foundry-provided generic template
       template: 'templates/generic/tab-navigation.hbs',
     },
-    programs: {
-      template: 'systems/swnr/templates/actor/cyberdeck/programs.hbs'
+    features: {
+      template: 'systems/swnr/templates/actor/features.hbs',
     },
-    // FRAGMENTS
-    programList: {
-      template: 'systems/swnr/templates/actor/cyberdeck/programs-list.hbs'
-    }
+    biography: {
+      template: 'systems/swnr/templates/actor/biography.hbs',
+    },
+    gear: {
+      template: 'systems/swnr/templates/actor/gear.hbs',
+    },
+    power: {
+      template: 'systems/swnr/templates/actor/powers.hbs',
+    },
+    effects: {
+      template: 'systems/swnr/templates/actor/effects.hbs',
+    },
   };
 
   /** @override */
   _configureRenderOptions(options) {
     super._configureRenderOptions(options);
-
-    //wsai adding to allow setting default tab
-    options.defaultTab = 'programs';
-
     // Not all parts always render
-    options.parts = ['header', 'tabs'];
+    options.parts = ['header', 'tabs', 'biography'];
     // Don't show the other tabs if only limited view
-    if (this.document.limited){ 
-      return;
-    }
+    if (this.document.limited) return;
     // Control which parts show based on document subtype
     switch (this.document.type) {
-      case 'cyberdeck':
-        options.parts.push('programs');
-        options.defaultTab = 'programs';
+      case 'mech':
+        //TODO
+        //options.parts.push('features', 'gear', 'powers', 'effects');
+        break;
+      case 'ship':
+        //TODO
+        //options.parts.push('gear', 'effects');
         break;
     }
   }
@@ -98,13 +99,10 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
       flags: this.actor.flags,
       // Adding a pointer to CONFIG.SWN
       config: CONFIG.SWN,
-      tabs: this._getTabs(options.parts, options.defaultTab),
+      tabs: this._getTabs(options.parts),
       // Necessary for formInput and formFields helpers
       fields: this.document.schema.fields,
       systemFields: this.document.system.schema.fields,
-      gameSettings: getGameSettings(),
-      headerWidget: headerFieldWidget.bind(this),
-      groupWidget: groupFieldWidget.bind(this),
     };
 
     // Offloading context prep to a helper function
@@ -115,12 +113,8 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
 
   /** @override */
   async _preparePartContext(partId, context) {
-    switch (partId) {
-      case 'programs':
-        context.tab = context.tabs[partId];
-        break;
-    }
-
+    // TODO copy from actor-sheet.mjs
+    console.log("TODO: Implement _preparePartContext");//TODO
     return context;
   }
 
@@ -130,11 +124,11 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
    * @returns {Record<string, Partial<ApplicationTab>>}
    * @protected
    */
-  _getTabs(parts, defaultTab = 'programs') {
+  _getTabs(parts) {
     // If you have sub-tabs this is necessary to change
     const tabGroup = 'primary';
     // Default tab for first time it's rendered this session
-    if (!this.tabGroups[tabGroup]) this.tabGroups[tabGroup] = defaultTab;
+    if (!this.tabGroups[tabGroup]) this.tabGroups[tabGroup] = 'biography';
     return parts.reduce((tabs, partId) => {
       const tab = {
         cssClass: 'sheet-body',
@@ -150,9 +144,25 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
         case 'header':
         case 'tabs':
           return tabs;
-        case 'programs':
-          tab.id = 'programs';
-          tab.label += 'Programs';
+        case 'biography':
+          tab.id = 'biography';
+          tab.label += 'Biography';
+          break;
+        case 'features':
+          tab.id = 'features';
+          tab.label += 'Features';
+          break;
+        case 'gear':
+          tab.id = 'gear';
+          tab.label += 'Gear';
+          break;
+        case 'spells':
+          tab.id = 'spells';
+          tab.label += 'Spells';
+          break;
+        case 'effects':
+          tab.id = 'effects';
+          tab.label += 'Effects';
           break;
       }
       if (this.tabGroups[tabGroup] === tab.id) tab.cssClass = 'active';
@@ -167,49 +177,11 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
    * @param {object} context The context object to mutate
    */
   _prepareItems(context) {
-    const hacker = this.actor.system.getHacker();
-
-    const programs = this.actor.items.filter(
-      (item) => item.type === "program"
-    );
-
-    const activePrograms = programs.filter(
-      (item) => item.system.type === "running"
-    );
-
-    const verbs = programs.filter(
-      (item) => item.system.type === "verb"
-    );
-
-    const subjects = programs.filter(
-      (item) => item.system.type === "subject"
-    );
-
-    const datafiles = programs.filter(
-      (item) => item.system.type === "datafile"
-    );
-
-    const access = {
-      value: 0,
-      max: 0,
-    };
-
-    if (hacker && (hacker.type == "character" || hacker.type == "npc")) {
-      access.max =
-        hacker.system.access.max + this.actor.system.bonusAccess;
-      access.value =
-        hacker.system.access.value + this.actor.system.bonusAccess;
-    }
-
-    return foundry.utils.mergeObject(context, {
-      itemTypes: this.actor.itemTypes,
-      activePrograms: activePrograms,
-      verbs: verbs,
-      subjects: subjects,
-      datafiles: datafiles,
-      hacker: hacker,
-      access: access
-    });
+    // Initialize containers.
+    // You can just use `this.document.itemTypes` instead
+    // if you don't need to subdivide a given type like
+    // this sheet does with spells
+    console.log("TODO: Implement _prepareItems");//TODO
   }
 
   /**
@@ -221,302 +193,12 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
    * @override
    */
   _onRender(context, options) {
-    super._onRender(context, options);
-
+    this.#dragDrop.forEach((d) => d.bind(this.element));
+    this.#disableOverrides();
     // You may want to add other special handling here
     // Foundry comes with a large number of utility classes, e.g. SearchFilter
     // That you may want to implement yourself.
   }
-
-  static async _onHackerRoll(event, target) {
-    event.preventDefault();
-    const actorId = this.actor.system.hackerId;
-    const crewActor = game.actors?.get(actorId);
-    if (!crewActor) {
-      ui.notifications?.error(`Hacker no longer exists`);
-      return;
-    }
-    const skills = crewActor.itemTypes.skill;
-    const isChar = crewActor.type == "character" ? true : false;
-    const dialogData = {
-      actor: crewActor,
-      skills: skills,
-      isChar,
-    };
-    const template = "systems/swnr/templates/dialogs/roll-skill-crew.hbs";
-    const html = await renderTemplate(template, dialogData);
-
-    const _rollForm = async (_event, button, html) => {
-      const rollMode = game.settings.get("core", "rollMode");
-      const dice = button.form.elements.dicepool.value;
-      const modifier = parseInt(
-        button.form.elements.modifier?.value
-      ) || 0;
-      const skillId = button.form.elements.skill?.value;
-      const skill = crewActor.getEmbeddedDocument(
-        "Item",
-        skillId
-      );
-      const useNPCSkillBonus = button.form.elements.useNPCSkillBonus?.checked
-        ? true : false;
-      const npcSkillBonus =
-        useNPCSkillBonus && crewActor.type == "npc"
-          ? crewActor.system.skillBonus
-          : 0;
-      const skillBonus = skill ? skill.system.rank : npcSkillBonus;
-      const statName = button.form.elements.stat
-        ?.value;
-      const stat = crewActor.system["stats"]?.[statName] || {
-        mod: 0,
-      };
-      const formula = `${dice} + @stat + @skillBonus + @modifier`;
-      const roll = new Roll(formula, {
-        skillBonus,
-        modifier,
-        stat: stat.mod,
-      });
-      const skillName = skill ? skill.name : "No Skill";
-      const statNameDisply = statName
-        ? game.i18n.localize("swnr.stat.short." + statName)
-        : "No Stat";
-      const title = `${game.i18n.localize(
-        "swnr.chat.skillCheck"
-      )}: ${statNameDisply}/${skillName}`;
-      await roll.roll({ async: true });
-      roll.toMessage(
-        {
-          speaker: { alias: crewActor.name },
-          flavor: title,
-        },
-        { rollMode }
-      );
-    };
-
-    const d = await foundry.applications.api.DialogV2.prompt({
-      window: {
-        title: game.i18n.format("swnr.dialog.skillRoll", {
-          actorName: crewActor?.name,
-        })
-      },
-      content: html,
-      modal: true,
-      rejectClose: false,
-      ok: {
-        label: game.i18n.localize("swnr.chat.roll"),
-        callback: _rollForm,
-      },
-    });
-  }
-
-  static async _onHackerDelete(event, target) {
-    const actorId = this.actor.system.hackerId;
-
-    if (actorId) {
-      await this.actor.update({
-        "system.hackerId": null
-      });
-      await this.actor.update({
-        "system.hacker": ""
-      });
-      let actor = game.actors?.get(actorId);
-      if (
-        actor &&
-        this.id &&
-        (actor.type === "character" || actor.type === "npc")
-      ) {
-        const cyberdeck = actor.system.cyberdecks;
-        const idx = cyberdeck.indexOf(this.id);
-        if (idx != -1) {
-          cyberdeck.splice(idx, 1);
-          await actor.update({
-            "system.cyberdecks": cyberdeck,
-          });
-        }
-      }
-      ui.notifications?.info(
-        `Removed hacker from ${this.actor.name}. Manually remove the cyberdeck from the hacker's sheet`
-      );
-    } else {
-      ui.notifications?.error("hacker not found");
-    }
-  }
-
-  static async _onActivateProgram(event, target) {
-    event.preventDefault();
-    if (this.actor.system.cpu.value < 1) {
-      ui.notifications?.error("Not enough CPU to activate program.");
-      return;
-    }
-    const sheetData = await this._prepareItems({});
-    const a = this.context;
-    let verbOptions = "";
-    let subjectOptions = "";
-    for (const verb of sheetData.verbs) {
-      verbOptions += `<option value="${verb.id}">${verb.name}</option>`;
-    }
-    for (const subject of sheetData.subjects) {
-      subjectOptions += `<option value="${subject.id}">${subject.name}</option>`;
-    }
-
-    if (!sheetData.verbs.length || !sheetData.subjects.length) {
-      ui.notifications?.error("No verbs or subjects found");
-      return;
-    }
-
-    const formContent = `<form>
-    <div class="form-group">
-      <label>Verb:</label>
-      <select name="verbId"
-        class="px-1.5 border border-gray-800 bg-gray-400 bg-opacity-75 placeholder-blue-800 placeholder-opacity-75 rounded-md">
-        ${verbOptions}
-      </select>
-      <label>Subject:</label>
-      <select name="subjectId"
-        class="px-1.5 border border-gray-800 bg-gray-400 bg-opacity-75 placeholder-blue-800 placeholder-opacity-75 rounded-md">
-        ${subjectOptions}
-      </select>
-    </div>
-    </form>`;
-
-    const _activateForm = async (html) => {
-      const form = html[0].querySelector("form");
-      const verbID = (form.querySelector('[name="verbId"]'))
-        ?.value;
-      const subID = (form.querySelector('[name="subjectId"]'))
-        ?.value;
-      if (!verbID || !subID) {
-        ui.notifications?.error("Verb or Subject not selected");
-        return;
-      }
-      const verb = sheetData.verbs.find((item) => item.id === verbID);
-      const subject = sheetData.subjects.find((item) => item.id === subID);
-      if (!verb || !subject) {
-        ui.notifications?.error("Verb or Subject not found");
-        return;
-      }
-
-      if (
-        verb.system.target &&
-          verb.system.target.indexOf(subject.system.target) === -1
-      ) {
-        ui.notifications?.error(
-          "Verb and Subject are incompatible (target and type does not match)"
-        );
-        return;
-      }
-      let skillCheckMod = 0;
-      if (verb.system.skillCheckMod) {
-        skillCheckMod += verb.system.skillCheckMod;
-      }
-      if (subject.system.skillCheckMod) {
-        skillCheckMod += subject.system.skillCheckMod;
-      }
-      const newProgram = {
-        name: `${verb.name} ${subject.name}`,
-        type: `program`,
-        img: verb.img,
-        system: {
-          type: "running",
-          cost: verb.system.cost,
-          accessCost: verb.system.accessCost,
-          useAffects: verb.system.useAffects,
-          selfTerminating: verb.system.selfTerminating,
-          skillCheckMod: skillCheckMod,
-        },
-      };
-
-      const docs = await this.actor.createEmbeddedDocuments(
-        "Item",
-        [newProgram],
-        {}
-      );
-      const program = docs[0];
-      if (!program) {
-        ui.notifications?.error("Failed to create program");
-        return;
-      }
-
-      // Consume access
-      if (sheetData.hacker) {
-        let access = 0;
-        if (sheetData.hacker.type == "character") {
-          access = sheetData.hacker.system.access.value;
-          access -= program.system.accessCost;
-        } else if (sheetData.hacker.type == "npc") {
-          access = sheetData.hacker.system.access.value;
-          access -= program.system.accessCost;
-        }
-        await sheetData.hacker.update({
-          "system.access.value": access,
-        });
-        if ((access + this.actor.system.bonusAccess) <= 0) {
-          ui.notifications?.info("Hacker has no access left");
-        }
-      }
-
-      // Roll skill / create button
-      program.roll();
-
-      if (program.system.selfTerminating) {
-        program.delete();
-      } else {
-        // await this.actor.update({
-        //   "data.cpu.value": this.actor.system.cpu.value - 1,
-        // });
-      }
-
-      this.render();
-    };
-
-    new Dialog({
-      title: game.i18n.localize("swnr.sheet.cyberdeck.run"),
-      content: formContent,
-      buttons: {
-        yes: {
-          icon: "<i class='fas fa-check'></i>",
-          label: `Run`,
-          callback: _activateForm,
-        },
-      },
-      default: "Run",
-    }).render(true);
-  }
-  
-
-  static async _onRefreshShielding(event, target) {
-    event.preventDefault();
-    const oldShielding = this.actor.system.health.value;
-    const maxShielding = this.actor.system.health.max;
-
-    ui.notifications?.info(
-      `Refreshing shielding from ${oldShielding} to ${maxShielding} (reminder 15 minutes to pass)`
-    );
-    await this.actor.update({
-      "system.health.value": maxShielding,
-    });
-  }
-
-  static async _onRefreshAccess(event, target) {
-    event.preventDefault();
-    const hacker = this.actor.system.getHacker();
-    if (hacker) {
-      const maxAccess = hacker.system.access.max;
-      const newAccessDisplay = maxAccess + this.actor.system.bonusAccess;
-      const oldAccess =
-        hacker.system.access.value + this.actor.system.bonusAccess;
-      ui.notifications?.info(
-        `Refreshing access from ${oldAccess} to ${newAccessDisplay} (reminder after 1 hour reprogramming, once per day)`
-      );
-      await hacker.update({
-        "system.access.value": maxAccess,
-      });
-      await this.actor.update({});
-    } else {
-      ui.notifications?.error("No hacker found");
-    }
-    this.render();
-  }
-
 
   /**************
    *
@@ -524,7 +206,158 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
    *
    **************/
 
+  /**
+   * Handle changing a Document's image.
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @returns {Promise}
+   * @protected
+   */
+  static async _onEditImage(event, target) {
+    const attr = target.dataset.edit;
+    const current = foundry.utils.getProperty(this.document, attr);
+    const { img } =
+      this.document.constructor.getDefaultArtwork?.(this.document.toObject()) ??
+      {};
+    const fp = new FilePicker({
+      current,
+      type: 'image',
+      redirectToRoot: img ? [img] : [],
+      callback: (path) => {
+        this.document.update({ [attr]: path });
+      },
+      top: this.position.top + 40,
+      left: this.position.left + 10,
+    });
+    return fp.browse();
+  }
+
+  /**
+   * Renders an embedded document's sheet
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _viewDoc(event, target) {
+    const doc = this._getEmbeddedDocument(target);
+    doc.sheet.render(true);
+  }
+
+  /**
+   * Handles item deletion
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _deleteDoc(event, target) {
+    const doc = this._getEmbeddedDocument(target);
+    await doc.delete();
+  }
+
+  /**
+   * Handle creating a new Owned Item or ActiveEffect for the actor using initial data defined in the HTML dataset
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @private
+   */
+  static async _createDoc(event, target) {
+    // Retrieve the configured document class for Item or ActiveEffect
+    const docCls = getDocumentClass(target.dataset.documentClass);
+    // Prepare the document creation data by initializing it a default name.
+    const docData = {
+      name: docCls.defaultName({
+        // defaultName handles an undefined type gracefully
+        type: target.dataset.type,
+        parent: this.actor,
+      }),
+    };
+    // Loop through the dataset and add it to our docData
+    for (const [dataKey, value] of Object.entries(target.dataset)) {
+      // These data attributes are reserved for the action handling
+      if (['action', 'documentClass'].includes(dataKey)) continue;
+      // Nested properties require dot notation in the HTML, e.g. anything with `system`
+      // An example exists in spells.hbs, with `data-system.spell-level`
+      // which turns into the dataKey 'system.spellLevel'
+      foundry.utils.setProperty(docData, dataKey, value);
+    }
+
+    // Finally, create the embedded document!
+    await docCls.create(docData, { parent: this.actor });
+  }
+
+  /**
+   * Determines effect parent to pass to helper
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @private
+   */
+  static async _toggleEffect(event, target) {
+    const effect = this._getEmbeddedDocument(target);
+    await effect.update({ disabled: !effect.disabled });
+  }
+
+  /**
+   * Handle clickable rolls.
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _onRoll(event, target) {
+    event.preventDefault();
+    const dataset = target.dataset;
+
+    // Handle item rolls.
+    switch (dataset.rollType) {
+      case 'item':
+        const item = this._getEmbeddedDocument(target);
+        if (item) return item.roll();
+    }
+
+    // Handle rolls that supply the formula directly.
+    if (dataset.roll) {
+      let label = dataset.label ? `[stat] ${dataset.label}` : '';
+      let roll = new Roll(dataset.roll, this.actor.getRollData());
+      await roll.toMessage({
+        speaker: ChatMessage.getSpeaker({ actor: this.actor }),
+        flavor: label,
+        rollMode: game.settings.get('core', 'rollMode'),
+      });
+      return roll;
+    }
+  }
+
   /** Helper Functions */
+
+  /**
+   * Fetches the embedded document representing the containing HTML element
+   *
+   * @param {HTMLElement} target    The element subject to search
+   * @returns {Item | ActiveEffect} The embedded Item or ActiveEffect
+   */
+  _getEmbeddedDocument(target) {
+    const docRow = target.closest('li[data-document-class]');
+    if (docRow.dataset.documentClass === 'Item') {
+      return this.actor.items.get(docRow.dataset.itemId);
+    } else if (docRow.dataset.documentClass === 'ActiveEffect') {
+      const parent =
+        docRow.dataset.parentId === this.actor.id
+          ? this.actor
+          : this.actor.items.get(docRow?.dataset.parentId);
+      return parent.effects.get(docRow?.dataset.effectId);
+    } else return console.warn('Could not find document class');
+  }
 
   /***************
    *
@@ -533,84 +366,323 @@ export class SWNCyberdeckSheet extends SWNBaseSheet {
    ***************/
 
   /**
-     * Handle dropping of an Actor data onto another Actor sheet
-     * @param {DragEvent} event            The concluding DragEvent which contains drop data
-     * @param {object} data                The data transfer extracted from the event
-     * @returns {Promise<object|boolean>}  A data object which describes the result of the drop, or false if the drop was
-     *                                     not permitted.
-     * @protected
-     */
-  async _onDropActor(_event, data) {
-    // Check ownership of the current actor
-    if (!this.actor.isOwner) return false;
+   * Define whether a user is able to begin a dragstart workflow for a given drag selector
+   * @param {string} selector       The candidate HTML selector for dragging
+   * @returns {boolean}             Can the current user drag this selector?
+   * @protected
+   */
+  _canDragStart(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
 
-    // Retrieve the Actor UUID from the data
-    const uuid = data.uuid;
+  /**
+   * Define whether a user is able to conclude a drag-and-drop workflow for a given drop selector
+   * @param {string} selector       The candidate HTML selector for the drop target
+   * @returns {boolean}             Can the current user drop on this selector?
+   * @protected
+   */
+  _canDragDrop(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
 
-    // Ensure the UUID exists
-    if (!uuid) {
-      console.error("No UUID found in the dropped data");
-      return false;
-    }
+  /**
+   * Callback actions which occur at the beginning of a drag start workflow.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  _onDragStart(event) {
+    const docRow = event.currentTarget.closest('li');
+    if ('link' in event.target.dataset) return;
 
-    if (this.actor.system.hackerId) {
-      //only one hacker allowed
-      ui.notifications?.error("Cyberdeck already has a hacker");
-      return;
-    }
+    // Chained operation
+    let dragData = this._getEmbeddedDocument(docRow)?.toDragData();
 
-    try {
-      // Retrieve the actor document using the UUID
-      const droppedActor = await fromUuid(uuid);
+    if (!dragData) return;
 
-      // Check if the retrieved document is an Actor
-      if (!droppedActor || !droppedActor instanceof Actor) {
-        console.error("The dropped data is not an actor.");
-        return false;
-      }
+    // Set data transfer
+    event.dataTransfer.setData('text/plain', JSON.stringify(dragData));
+  }
 
-      // Check if the refreived document is a Character or NPC.
-      if (droppedActor.type != 'character' && droppedActor.type != 'npc') {
-        ui.notifications?.error("The dropped/added hacker is not a character or an npc.");
-        return false;
-      }
-      const droppedActorId = droppedActor.id;
-      await this.actor.update({ "system.hackerId": droppedActorId });
-      const cyberdecks = droppedActor.system.cyberdecks;
-      if (this.id && cyberdecks.indexOf(this.id) === -1) {
-        cyberdecks.push(this.actor.id);
-        await droppedActor.update({
-          "system.cyberdecks": cyberdecks,
-        });
-      }
+  /**
+   * Callback actions which occur when a dragged element is over a drop target.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  _onDragOver(event) {}
 
-      const itemName = this.actor.name;
-      droppedActor.createEmbeddedDocuments(
-        "Item",
-        [
-          {
-            name: itemName,
-            type: "item",
-            img: "systems/swnr/assets/icons/cyberdeck.png",
-            system: {
-              encumbrance: this.actor.system.encumberance,
-              quantity: 1,
-              cost: this.actor.system.cost,
-            },
-          },
-        ],
-        {}
-      );
-      ui.notifications?.info(
-        `Created a cyberdeck item "${itemName}" on ${droppedActor.name}'s sheet`
-      );
+  /**
+   * Callback actions which occur when a dragged element is dropped on a target.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  async _onDrop(event) {
+    const data = TextEditor.getDragEventData(event);
+    const actor = this.actor;
+    const allowed = Hooks.call('dropActorSheetData', actor, this, data);
+    if (allowed === false) return;
 
-    } catch (err) {
-      console.error("Error retrieving actor from UUID:", err);
+    // Handle different data types
+    switch (data.type) {
+      case 'ActiveEffect':
+        return this._onDropActiveEffect(event, data);
+      case 'Actor':
+        return this._onDropActor(event, data);
+      case 'Item':
+        return this._onDropItem(event, data);
+      case 'Folder':
+        return this._onDropFolder(event, data);
     }
   }
 
+  /**
+   * Handle the dropping of ActiveEffect data onto an Actor Sheet
+   * @param {DragEvent} event                  The concluding DragEvent which contains drop data
+   * @param {object} data                      The data transfer extracted from the event
+   * @returns {Promise<ActiveEffect|boolean>}  The created ActiveEffect object or false if it couldn't be created.
+   * @protected
+   */
+  async _onDropActiveEffect(event, data) {
+    const aeCls = getDocumentClass('ActiveEffect');
+    const effect = await aeCls.fromDropData(data);
+    if (!this.actor.isOwner || !effect) return false;
+    if (effect.target === this.actor)
+      return this._onSortActiveEffect(event, effect);
+    return aeCls.create(effect, { parent: this.actor });
+  }
 
-  // All in base
+  /**
+   * Handle a drop event for an existing embedded Active Effect to sort that Active Effect relative to its siblings
+   *
+   * @param {DragEvent} event
+   * @param {ActiveEffect} effect
+   */
+  async _onSortActiveEffect(event, effect) {
+    /** @type {HTMLElement} */
+    const dropTarget = event.target.closest('[data-effect-id]');
+    if (!dropTarget) return;
+    const target = this._getEmbeddedDocument(dropTarget);
 
+    // Don't sort on yourself
+    if (effect.uuid === target.uuid) return;
+
+    // Identify sibling items based on adjacent HTML elements
+    const siblings = [];
+    for (const el of dropTarget.parentElement.children) {
+      const siblingId = el.dataset.effectId;
+      const parentId = el.dataset.parentId;
+      if (
+        siblingId &&
+        parentId &&
+        (siblingId !== effect.id || parentId !== effect.parent.id)
+      )
+        siblings.push(this._getEmbeddedDocument(el));
+    }
+
+    // Perform the sort
+    const sortUpdates = SortingHelpers.performIntegerSort(effect, {
+      target,
+      siblings,
+    });
+
+    // Split the updates up by parent document
+    const directUpdates = [];
+
+    const grandchildUpdateData = sortUpdates.reduce((items, u) => {
+      const parentId = u.target.parent.id;
+      const update = { _id: u.target.id, ...u.update };
+      if (parentId === this.actor.id) {
+        directUpdates.push(update);
+        return items;
+      }
+      if (items[parentId]) items[parentId].push(update);
+      else items[parentId] = [update];
+      return items;
+    }, {});
+
+    // Effects-on-items updates
+    for (const [itemId, updates] of Object.entries(grandchildUpdateData)) {
+      await this.actor.items
+        .get(itemId)
+        .updateEmbeddedDocuments('ActiveEffect', updates);
+    }
+
+    // Update on the main actor
+    return this.actor.updateEmbeddedDocuments('ActiveEffect', directUpdates);
+  }
+
+  /**
+   * Handle dropping of an Actor data onto another Actor sheet
+   * @param {DragEvent} event            The concluding DragEvent which contains drop data
+   * @param {object} data                The data transfer extracted from the event
+   * @returns {Promise<object|boolean>}  A data object which describes the result of the drop, or false if the drop was
+   *                                     not permitted.
+   * @protected
+   */
+  async _onDropActor(event, data) {
+    if (!this.actor.isOwner) return false;
+  }
+
+  /* -------------------------------------------- */
+
+  /**
+   * Handle dropping of an item reference or item data onto an Actor Sheet
+   * @param {DragEvent} event            The concluding DragEvent which contains drop data
+   * @param {object} data                The data transfer extracted from the event
+   * @returns {Promise<Item[]|boolean>}  The created or updated Item instances, or false if the drop was not permitted.
+   * @protected
+   */
+  async _onDropItem(event, data) {
+    if (!this.actor.isOwner) return false;
+    const item = await Item.implementation.fromDropData(data);
+
+    // Handle item sorting within the same Actor
+    if (this.actor.uuid === item.parent?.uuid)
+      return this._onSortItem(event, item);
+
+    // Create the owned item
+    return this._onDropItemCreate(item, event);
+  }
+
+  /**
+   * Handle dropping of a Folder on an Actor Sheet.
+   * The core sheet currently supports dropping a Folder of Items to create all items as owned items.
+   * @param {DragEvent} event     The concluding DragEvent which contains drop data
+   * @param {object} data         The data transfer extracted from the event
+   * @returns {Promise<Item[]>}
+   * @protected
+   */
+  async _onDropFolder(event, data) {
+    if (!this.actor.isOwner) return [];
+    const folder = await Folder.implementation.fromDropData(data);
+    if (folder.type !== 'Item') return [];
+    const droppedItemData = await Promise.all(
+      folder.contents.map(async (item) => {
+        if (!(document instanceof Item)) item = await fromUuid(item.uuid);
+        return item;
+      })
+    );
+    return this._onDropItemCreate(droppedItemData, event);
+  }
+
+  /**
+   * Handle the final creation of dropped Item data on the Actor.
+   * This method is factored out to allow downstream classes the opportunity to override item creation behavior.
+   * @param {object[]|object} itemData      The item data requested for creation
+   * @param {DragEvent} event               The concluding DragEvent which provided the drop data
+   * @returns {Promise<Item[]>}
+   * @private
+   */
+  async _onDropItemCreate(itemData, event) {
+    itemData = itemData instanceof Array ? itemData : [itemData];
+    return this.actor.createEmbeddedDocuments('Item', itemData);
+  }
+
+  /**
+   * Handle a drop event for an existing embedded Item to sort that Item relative to its siblings
+   * @param {Event} event
+   * @param {Item} item
+   * @private
+   */
+  _onSortItem(event, item) {
+    // Get the drag source and drop target
+    const items = this.actor.items;
+    const dropTarget = event.target.closest('[data-item-id]');
+    if (!dropTarget) return;
+    const target = items.get(dropTarget.dataset.itemId);
+
+    // Don't sort on yourself
+    if (item.id === target.id) return;
+
+    // Identify sibling items based on adjacent HTML elements
+    const siblings = [];
+    for (let el of dropTarget.parentElement.children) {
+      const siblingId = el.dataset.itemId;
+      if (siblingId && siblingId !== item.id)
+        siblings.push(items.get(el.dataset.itemId));
+    }
+
+    // Perform the sort
+    const sortUpdates = SortingHelpers.performIntegerSort(item, {
+      target,
+      siblings,
+    });
+    const updateData = sortUpdates.map((u) => {
+      const update = u.update;
+      update._id = u.target._id;
+      return update;
+    });
+
+    // Perform the update
+    return this.actor.updateEmbeddedDocuments('Item', updateData);
+  }
+
+  /** The following pieces set up drag handling and are unlikely to need modification  */
+
+  /**
+   * Returns an array of DragDrop instances
+   * @type {DragDrop[]}
+   */
+  get dragDrop() {
+    return this.#dragDrop;
+  }
+
+  // This is marked as private because there's no real need
+  // for subclasses or external hooks to mess with it directly
+  #dragDrop;
+
+  /**
+   * Create drag-and-drop workflow handlers for this Application
+   * @returns {DragDrop[]}     An array of DragDrop handlers
+   * @private
+   */
+  #createDragDropHandlers() {
+    return this.options.dragDrop.map((d) => {
+      d.permissions = {
+        dragstart: this._canDragStart.bind(this),
+        drop: this._canDragDrop.bind(this),
+      };
+      d.callbacks = {
+        dragstart: this._onDragStart.bind(this),
+        dragover: this._onDragOver.bind(this),
+        drop: this._onDrop.bind(this),
+      };
+      return new DragDrop(d);
+    });
+  }
+
+  /********************
+   *
+   * Actor Override Handling
+   *
+   ********************/
+
+  /**
+   * Submit a document update based on the processed form data.
+   * @param {SubmitEvent} event                   The originating form submission event
+   * @param {HTMLFormElement} form                The form element that was submitted
+   * @param {object} submitData                   Processed and validated form data to be used for a document update
+   * @returns {Promise<void>}
+   * @protected
+   * @override
+   */
+  async _processSubmitData(event, form, submitData) {
+    const overrides = foundry.utils.flattenObject(this.actor.overrides);
+    for (let k of Object.keys(overrides)) delete submitData[k];
+    await this.document.update(submitData);
+  }
+
+  /**
+   * Disables inputs subject to active effects
+   */
+  #disableOverrides() {
+    const flatOverrides = foundry.utils.flattenObject(this.actor.overrides);
+    for (const override of Object.keys(flatOverrides)) {
+      const input = this.element.querySelector(`[name="${override}"]`);
+      if (input) {
+        input.disabled = true;
+      }
+    }
+  }
 }
