+++ module/sheets/vehicle-sheet.mjs
@@ -1,7 +1,7 @@
 import { prepareActiveEffectCategories } from '../helpers/effects.mjs';
 import { getGameSettings } from '../helpers/register-settings.mjs';
 import { headerFieldWidget, groupFieldWidget } from '../helpers/handlebar.mjs';
-import { SWNBaseSheet } from './base-sheet.mjs';
+
 
 const { api, sheets } = foundry.applications;
 
@@ -9,9 +9,12 @@ const { api, sheets } = foundry.applications;
  * Extend the basic ActorSheet with some very simple modifications
  * @extends {ActorSheetV2}
  */
-export class SWNVehicleSheet extends SWNBaseSheet {
+export class SWNVehicleSheet extends api.HandlebarsApplicationMixin(
+  sheets.ActorSheetV2
+) {
   constructor(options = {}) {
     super(options);
+    this.#dragDrop = this.#createDragDropHandlers();
   }
 
   /** @override */
@@ -41,11 +44,7 @@ export class SWNVehicleSheet extends SWNBaseSheet {
       payMaintenance: this._onMaintenance,
       creditsChange: this._onAddCurrency,
       resourceDelete: this._onResourceDelete,
-      resourceCreate: this._onResourceCreate,
-      crewDelete: this._onCrewDelete,
-      crewRoll: this._onCrewRoll,
-      crewShow: this._onCrewShow,
-      toggle: this._onToggleVehicleStatus,
+      resourceCreate: this._onResourceCreate
     },
     // Custom property that's merged into `this.options`
     dragDrop: [{ dragSelector: '[data-drag]', dropSelector: null }],
@@ -83,30 +82,6 @@ export class SWNVehicleSheet extends SWNBaseSheet {
     },
     mech: {
       template: 'systems/swnr/templates/actor/vehicle/mech.hbs',
-    },
-    shipCrew: {
-      template: 'systems/swnr/templates/actor/vehicle/ship-crew.hbs',
-    },
-    crewList: {
-      template: 'systems/swnr/templates/actor/fragments/crew-list.hbs',
-    },
-    shipCombat: {
-      template: 'systems/swnr/templates/actor/vehicle/ship-combat.hbs',
-    },
-    shipCargo: {
-      template: 'systems/swnr/templates/actor/vehicle/ship-cargo.hbs',
-    },
-    carriedGear: {
-      template: 'systems/swnr/templates/actor/fragments/compact-carried-list.hbs'
-    },
-    vehicleWeapons: {
-      template: 'systems/swnr/templates/actor/fragments/vehicle-weapon-table.hbs'
-    },
-    vehicleFittings: {
-      template: 'systems/swnr/templates/actor/fragments/vehicle-fitting-table.hbs'
-    },
-    vehicleDefense: {
-      template: 'systems/swnr/templates/actor/fragments/vehicle-defense-table.hbs'
     }
   };
 
@@ -126,8 +101,8 @@ export class SWNVehicleSheet extends SWNBaseSheet {
         options.defaultTab = 'mech';
         break;
       case 'ship':
-        options.parts.push('shipCombat','ship','shipCrew','shipCargo','effects');
-        options.defaultTab = 'shipCombat';
+        options.parts.push('ship','effects');
+        options.defaultTab = 'ship';
         break;
       case 'drone':
         options.parts.push('drone','effects');
@@ -144,15 +119,6 @@ export class SWNVehicleSheet extends SWNBaseSheet {
 
   /** @override */
   async _prepareContext(options) {
-    let crewArray = [];
-    if (this.actor.system.crewMembers) {
-      for (let crew of this.actor.system.crewMembers) {
-        let actor = game.actors.get(crew);
-        if (actor) {
-          crewArray.push(actor);
-        } 
-      }
-    }
     // Output initialization
     const context = {
       // Validates both permissions and compendium status
@@ -166,15 +132,13 @@ export class SWNVehicleSheet extends SWNBaseSheet {
       flags: this.actor.flags,
       // Adding a pointer to CONFIG.SWN
       config: CONFIG.SWN,
-      tabs: this._getTabs(options.parts, options.defaultTab),
+      tabs: this._getTabs(options.parts),
       // Necessary for formInput and formFields helpers
       fields: this.document.schema.fields,
       systemFields: this.document.system.schema.fields,
       gameSettings: getGameSettings(),
       headerWidget: headerFieldWidget.bind(this),
       groupWidget: groupFieldWidget.bind(this),
-      crewArray: crewArray,
-      actions: CONFIG.SWN.shipActions
     };
 
     // Offloading context prep to a helper function
@@ -189,9 +153,6 @@ export class SWNVehicleSheet extends SWNBaseSheet {
     console.log("TODO: Implement _preparePartContext");//TODO
     switch (partId) {
       case 'ship':
-      case 'shipCrew':
-      case 'shipCombat':
-      case 'shipCargo':
       case 'vehicle':
       case 'drone':
       case 'mech':
@@ -268,18 +229,6 @@ export class SWNVehicleSheet extends SWNBaseSheet {
           tab.id = 'ship';
           tab.label += 'Ship';
           break;
-        case 'shipCrew':
-          tab.id = 'shipCrew';
-          tab.label += 'Crew';
-          break;
-        case 'shipCargo':
-          tab.id = 'shipCargo';
-          tab.label += 'Cargo';
-          break;
-        case 'shipCombat':
-          tab.id = 'shipCombat';
-          tab.label += 'Combat';
-          break;
         case 'vehicle':
           tab.id = 'vehicle';
           tab.label += 'Vehicle';
@@ -321,13 +270,12 @@ export class SWNVehicleSheet extends SWNBaseSheet {
    * @override
    */
   _onRender(context, options) {
-    super._onRender(context, options);
-
+    this.#dragDrop.forEach((d) => d.bind(this.element));
+    this.#disableOverrides();
     // You may want to add other special handling here
     // Foundry comes with a large number of utility classes, e.g. SearchFilter
     // That you may want to implement yourself.
 
-    
     this.element.querySelectorAll("[name='shipActions']").forEach((d) => 
       d.addEventListener('change', this._onShipAction.bind(this)));
     this.element.querySelectorAll("[name='data.shipHullType']").forEach((d) => 
@@ -336,12 +284,317 @@ export class SWNVehicleSheet extends SWNBaseSheet {
       d.addEventListener('change', this._onResourceName.bind(this)));
   }
 
+  /**************
+   *
+   *   ACTIONS
+   *
+   **************/
 
+  /**
+   * Handle changing a Document's image.
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @returns {Promise}
+   * @protected
+   */
+  static async _onEditImage(event, target) {
+    const attr = target.dataset.edit;
+    const current = foundry.utils.getProperty(this.document, attr);
+    const { img } =
+      this.document.constructor.getDefaultArtwork?.(this.document.toObject()) ??
+      {};
+    const fp = new FilePicker({
+      current,
+      type: 'image',
+      redirectToRoot: img ? [img] : [],
+      callback: (path) => {
+        this.document.update({ [attr]: path });
+      },
+      top: this.position.top + 40,
+      left: this.position.left + 10,
+    });
+    return fp.browse();
+  }
+
+  /**
+   * Renders an embedded document's sheet
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _viewDoc(event, target) {
+    const doc = this._getEmbeddedDocument(target);
+    doc.sheet.render(true);
+  }
+
+  /**
+   * Handles item deletion
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _deleteDoc(event, target) {
+    const doc = this._getEmbeddedDocument(target);
+    await doc.delete();
+  }
+
+  /**
+   * Handle creating a new Owned Item or ActiveEffect for the actor using initial data defined in the HTML dataset
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @private
+   */
+  static async _createDoc(event, target) {
+    // Retrieve the configured document class for Item or ActiveEffect
+    const docCls = getDocumentClass(target.dataset.documentClass);
+    // Prepare the document creation data by initializing it a default name.
+    const docData = {
+      name: docCls.defaultName({
+        // defaultName handles an undefined type gracefully
+        type: target.dataset.type,
+        parent: this.actor,
+      }),
+    };
+    // Loop through the dataset and add it to our docData
+    for (const [dataKey, value] of Object.entries(target.dataset)) {
+      // These data attributes are reserved for the action handling
+      if (['action', 'documentClass'].includes(dataKey)) continue;
+      // Nested properties require dot notation in the HTML, e.g. anything with `system`
+      // An example exists in spells.hbs, with `data-system.spell-level`
+      // which turns into the dataKey 'system.spellLevel'
+      foundry.utils.setProperty(docData, dataKey, value);
+    }
+
+    // Finally, create the embedded document!
+    await docCls.create(docData, { parent: this.actor });
+  }
+
+  /**
+   * Determines effect parent to pass to helper
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @private
+   */
+  static async _toggleEffect(event, target) {
+    const effect = this._getEmbeddedDocument(target);
+    await effect.update({ disabled: !effect.disabled });
+  }
+
+  /**
+   * Handle clickable rolls.
+   *
+   * @this SWNActorSheet
+   * @param {PointerEvent} event   The originating click event
+   * @param {HTMLElement} target   The capturing HTML element which defined a [data-action]
+   * @protected
+   */
+  static async _onRoll(event, target) {
+    event.preventDefault();
+    const dataset = target.dataset;
+
+    // Handle item rolls.
+    switch (dataset.rollType) {
+      case 'item':
+        const item = this._getEmbeddedDocument(target);
+        if (item) return item.roll();
+    }
+
+    // Handle rolls that supply the formula directly.
+    if (dataset.roll) {
+      let label = dataset.label ? `[stat] ${dataset.label}` : '';
+      let roll = new Roll(dataset.roll, this.actor.getRollData());
+      await roll.toMessage({
+        speaker: ChatMessage.getSpeaker({ actor: this.actor }),
+        flavor: label,
+        rollMode: game.settings.get('core', 'rollMode'),
+      });
+      return roll;
+    }
+  }
 
   /** Helper Functions */
 
+  /**
+   * Fetches the embedded document representing the containing HTML element
+   *
+   * @param {HTMLElement} target    The element subject to search
+   * @returns {Item | ActiveEffect} The embedded Item or ActiveEffect
+   */
+  _getEmbeddedDocument(target) {
+    const docRow = target.closest('li[data-document-class]');
+    if (docRow.dataset.documentClass === 'Item') {
+      return this.actor.items.get(docRow.dataset.itemId);
+    } else if (docRow.dataset.documentClass === 'ActiveEffect') {
+      const parent =
+        docRow.dataset.parentId === this.actor.id
+          ? this.actor
+          : this.actor.items.get(docRow?.dataset.parentId);
+      return parent.effects.get(docRow?.dataset.effectId);
+    } else return console.warn('Could not find document class');
+  }
 
+  /***************
+   *
+   * Drag and Drop
+   *
+   ***************/
 
+  /**
+   * Define whether a user is able to begin a dragstart workflow for a given drag selector
+   * @param {string} selector       The candidate HTML selector for dragging
+   * @returns {boolean}             Can the current user drag this selector?
+   * @protected
+   */
+  _canDragStart(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
+
+  /**
+   * Define whether a user is able to conclude a drag-and-drop workflow for a given drop selector
+   * @param {string} selector       The candidate HTML selector for the drop target
+   * @returns {boolean}             Can the current user drop on this selector?
+   * @protected
+   */
+  _canDragDrop(selector) {
+    // game.user fetches the current user
+    return this.isEditable;
+  }
+
+  /**
+   * Callback actions which occur at the beginning of a drag start workflow.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  _onDragStart(event) {
+    const docRow = event.currentTarget.closest('li');
+    if ('link' in event.target.dataset) return;
+
+    // Chained operation
+    let dragData = this._getEmbeddedDocument(docRow)?.toDragData();
+
+    if (!dragData) return;
+
+    // Set data transfer
+    event.dataTransfer.setData('text/plain', JSON.stringify(dragData));
+  }
+
+  /**
+   * Callback actions which occur when a dragged element is over a drop target.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  _onDragOver(event, target) {}
+
+  /**
+   * Callback actions which occur when a dragged element is dropped on a target.
+   * @param {DragEvent} event       The originating DragEvent
+   * @protected
+   */
+  async _onDrop(event, target) {
+    const data = TextEditor.getDragEventData(event);
+    const actor = this.actor;
+    const allowed = Hooks.call('dropActorSheetData', actor, this, data);
+    if (allowed === false) return;
+
+    // Handle different data types
+    switch (data.type) {
+      case 'ActiveEffect':
+        return this._onDropActiveEffect(event, data);
+      case 'Actor':
+        return this._onDropActor(event, data);
+      case 'Item':
+        return this._onDropItem(event, data);
+      case 'Folder':
+        return this._onDropFolder(event, data);
+    }
+  }
+
+  /**
+   * Handle the dropping of ActiveEffect data onto an Actor Sheet
+   * @param {DragEvent} event                  The concluding DragEvent which contains drop data
+   * @param {object} data                      The data transfer extracted from the event
+   * @returns {Promise<ActiveEffect|boolean>}  The created ActiveEffect object or false if it couldn't be created.
+   * @protected
+   */
+  async _onDropActiveEffect(event, data) {
+    const aeCls = getDocumentClass('ActiveEffect');
+    const effect = await aeCls.fromDropData(data);
+    if (!this.actor.isOwner || !effect) return false;
+    if (effect.target === this.actor)
+      return this._onSortActiveEffect(event, effect);
+    return aeCls.create(effect, { parent: this.actor });
+  }
+
+  /**
+   * Handle a drop event for an existing embedded Active Effect to sort that Active Effect relative to its siblings
+   *
+   * @param {DragEvent} event
+   * @param {ActiveEffect} effect
+   */
+  async _onSortActiveEffect(event, effect) {
+    /** @type {HTMLElement} */
+    const dropTarget = event.target.closest('[data-effect-id]');
+    if (!dropTarget) return;
+    const target = this._getEmbeddedDocument(dropTarget);
+
+    // Don't sort on yourself
+    if (effect.uuid === target.uuid) return;
+
+    // Identify sibling items based on adjacent HTML elements
+    const siblings = [];
+    for (const el of dropTarget.parentElement.children) {
+      const siblingId = el.dataset.effectId;
+      const parentId = el.dataset.parentId;
+      if (
+        siblingId &&
+        parentId &&
+        (siblingId !== effect.id || parentId !== effect.parent.id)
+      )
+        siblings.push(this._getEmbeddedDocument(el));
+    }
+
+    // Perform the sort
+    const sortUpdates = SortingHelpers.performIntegerSort(effect, {
+      target,
+      siblings,
+    });
+
+    // Split the updates up by parent document
+    const directUpdates = [];
+
+    const grandchildUpdateData = sortUpdates.reduce((items, u) => {
+      const parentId = u.target.parent.id;
+      const update = { _id: u.target.id, ...u.update };
+      if (parentId === this.actor.id) {
+        directUpdates.push(update);
+        return items;
+      }
+      if (items[parentId]) items[parentId].push(update);
+      else items[parentId] = [update];
+      return items;
+    }, {});
+
+    // Effects-on-items updates
+    for (const [itemId, updates] of Object.entries(grandchildUpdateData)) {
+      await this.actor.items
+        .get(itemId)
+        .updateEmbeddedDocuments('ActiveEffect', updates);
+    }
+
+    // Update on the main actor
+    return this.actor.updateEmbeddedDocuments('ActiveEffect', directUpdates);
+  }
 
   /**
    * Handle dropping of an Actor data onto another Actor sheet
@@ -353,35 +606,135 @@ export class SWNVehicleSheet extends SWNBaseSheet {
    */
   async _onDropActor(event, data) {
     if (!this.actor.isOwner) return false;
-    //dropped actor
-    let actor = fromUuidSync(data.uuid);
-    // Only allow dropping of npc or character actors
-    if (actor.type !== 'npc' && actor.type !== 'character') {
-      ui.notifications.error("Only NPC or Character actors can be dropped onto vehicle.");
-      return false;
-    }
-    const actorId = actor.id;
-    if (this.actor.type === 'ship' || this.actor.type === 'vehicle') {
-      //Multi-crew vehicle
-      let crewMembers = this.actor.system.crewMembers;
-      if (crewMembers.indexOf(actorId) == -1) {
-        crewMembers.push(actorId);
-        let crewCount = this.actor.system.crew.current + 1;
-        await this.actor.update(
-          { system: {
-            crewMembers: crewMembers,
-            crew: { current: crewCount }
-          }
-        });
-      }
-    } else {
-      //Single pilot vehicle
-      ui.notifications.error("TODO");
-    }
   }
 
   /* -------------------------------------------- */
 
+  /**
+   * Handle dropping of an item reference or item data onto an Actor Sheet
+   * @param {DragEvent} event            The concluding DragEvent which contains drop data
+   * @param {object} data                The data transfer extracted from the event
+   * @returns {Promise<Item[]|boolean>}  The created or updated Item instances, or false if the drop was not permitted.
+   * @protected
+   */
+  async _onDropItem(event, data) {
+    if (!this.actor.isOwner) return false;
+    const item = await Item.implementation.fromDropData(data);
+
+    // Handle item sorting within the same Actor
+    if (this.actor.uuid === item.parent?.uuid)
+      return this._onSortItem(event, item);
+
+    // Create the owned item
+    return this._onDropItemCreate(item, event);
+  }
+
+  /**
+   * Handle dropping of a Folder on an Actor Sheet.
+   * The core sheet currently supports dropping a Folder of Items to create all items as owned items.
+   * @param {DragEvent} event     The concluding DragEvent which contains drop data
+   * @param {object} data         The data transfer extracted from the event
+   * @returns {Promise<Item[]>}
+   * @protected
+   */
+  async _onDropFolder(event, data) {
+    if (!this.actor.isOwner) return [];
+    const folder = await Folder.implementation.fromDropData(data);
+    if (folder.type !== 'Item') return [];
+    const droppedItemData = await Promise.all(
+      folder.contents.map(async (item) => {
+        if (!(document instanceof Item)) item = await fromUuid(item.uuid);
+        return item;
+      })
+    );
+    return this._onDropItemCreate(droppedItemData, event);
+  }
+
+  /**
+   * Handle the final creation of dropped Item data on the Actor.
+   * This method is factored out to allow downstream classes the opportunity to override item creation behavior.
+   * @param {object[]|object} itemData      The item data requested for creation
+   * @param {DragEvent} event               The concluding DragEvent which provided the drop data
+   * @returns {Promise<Item[]>}
+   * @private
+   */
+  async _onDropItemCreate(itemData, event) {
+    itemData = itemData instanceof Array ? itemData : [itemData];
+    return this.actor.createEmbeddedDocuments('Item', itemData);
+  }
+
+  /**
+   * Handle a drop event for an existing embedded Item to sort that Item relative to its siblings
+   * @param {Event} event
+   * @param {Item} item
+   * @private
+   */
+  _onSortItem(event, item) {
+    // Get the drag source and drop target
+    const items = this.actor.items;
+    const dropTarget = event.target.closest('[data-item-id]');
+    if (!dropTarget) return;
+    const target = items.get(dropTarget.dataset.itemId);
+
+    // Don't sort on yourself
+    if (item.id === target.id) return;
+
+    // Identify sibling items based on adjacent HTML elements
+    const siblings = [];
+    for (let el of dropTarget.parentElement.children) {
+      const siblingId = el.dataset.itemId;
+      if (siblingId && siblingId !== item.id)
+        siblings.push(items.get(el.dataset.itemId));
+    }
+
+    // Perform the sort
+    const sortUpdates = SortingHelpers.performIntegerSort(item, {
+      target,
+      siblings,
+    });
+    const updateData = sortUpdates.map((u) => {
+      const update = u.update;
+      update._id = u.target._id;
+      return update;
+    });
+
+    // Perform the update
+    return this.actor.updateEmbeddedDocuments('Item', updateData);
+  }
+
+  /** The following pieces set up drag handling and are unlikely to need modification  */
+
+  /**
+   * Returns an array of DragDrop instances
+   * @type {DragDrop[]}
+   */
+  get dragDrop() {
+    return this.#dragDrop;
+  }
+
+  // This is marked as private because there's no real need
+  // for subclasses or external hooks to mess with it directly
+  #dragDrop;
+
+  /**
+   * Create drag-and-drop workflow handlers for this Application
+   * @returns {DragDrop[]}     An array of DragDrop handlers
+   * @private
+   */
+  #createDragDropHandlers() {
+    return this.options.dragDrop.map((d) => {
+      d.permissions = {
+        dragstart: this._canDragStart.bind(this),
+        drop: this._canDragDrop.bind(this),
+      };
+      d.callbacks = {
+        dragstart: this._onDragStart.bind(this),
+        dragover: this._onDragOver.bind(this),
+        drop: this._onDrop.bind(this),
+      };
+      return new DragDrop(d);
+    });
+  }
 
   // Add Currency to selected type
 static async _onAddCurrency(event, target) {
@@ -393,10 +746,10 @@ static async _onAddCurrency(event, target) {
   const currencyType = target.dataset.creditType
 
   // load html variable data for dialog
-  const template = "systems/swnr/templates/dialogs/add-currency.hbs";
+  const template = "systems/swnr/templates/dialogs/add-currency.html";
   const data = {};
   const html = await renderTemplate(template, data);
-
+  this.popUpDialog?.close();
   
   // show a dialog prompting for amount of change to add
   const amount = await new Promise((resolve) => {
@@ -433,19 +786,19 @@ static async _onAddCurrency(event, target) {
   } else {
     // this is our valid input scenario
     await this.actor.update({
-      system: {
+      data: {
         [currencyType]: this.actor.system[currencyType] + parseInt(await amount)
       }
     });
   }
 }
 
-async _onResourceName(event, target) {
+static async _onResourceName(event, target) {
   event.preventDefault();
   event.stopPropagation();
-  const value = event.target?.value;
-  const resourceType = event.target.dataset.rlType
-  const idx = event.target.dataset.rlIdx;
+  const value = target?.value;
+  const resourceType = target.dataset.rlType
+  const idx = $(event.currentTarget).parents(".item").data("rlIdx");
   const resourceList = duplicate(this.actor.system.cargoCarried);
   resourceList[idx][resourceType] = value;
   await this.actor.update({ "system.cargoCarried": resourceList });
@@ -454,7 +807,7 @@ async _onResourceName(event, target) {
 static async _onResourceDelete(event, target) {
   event.preventDefault();
   event.stopPropagation();
-  const idx = target.dataset.rlIdx;
+  const idx = $(event.currentTarget).parents(".item").data("rlIdx");
   const resourceList = duplicate(this.actor.system.cargoCarried);
   resourceList.splice(idx, 1);
   await this.actor.update({ "system.cargoCarried": resourceList });
@@ -462,6 +815,7 @@ static async _onResourceDelete(event, target) {
 
 static async _onResourceCreate(event, target) {
   event.preventDefault();
+  // console.log("Changing HP Max", this.actor);
   let resourceList = this.actor.system.cargoCarried;
   if (!resourceList) {
     resourceList = [];
@@ -469,158 +823,6 @@ static async _onResourceCreate(event, target) {
   resourceList.push({ name: "Cargo X", value: 0, max: 1 });
   await this.actor.update({
     "system.cargoCarried": resourceList,
-  });
-}
-
-static async _onCrewDelete(event, target) {
-  event.preventDefault();
-  const actorId = target.dataset.crewId;
-  let crewMembers = this.actor.system.crewMembers;
-  //Only remove if there
-  const idx = crewMembers.indexOf(actorId);
-  if (idx == -1) {
-    ui.notifications?.error("Crew member not found");
-  } else {
-    crewMembers.splice(idx, 1);
-    let crew = this.actor.system.crew.current;
-    crew -= 1;
-    if (this.actor.system.roles) {
-      let roles = {...this.actor.system.roles};
-      if (roles.captain == actorId) {
-        roles.captain = null;
-      }
-      if (roles.comms == actorId) {
-        roles.comms = null;
-      }
-      if (roles.engineering == actorId) {
-        roles.engineering = null;
-      }
-      if (roles.gunnery == actorId) {
-        roles.gunnery = null;
-      }
-      if (roles.bridge == actorId) {
-        roles.bridge = null;
-      }
-      await this.actor.update({
-        "system.crew.current": crew,
-        "system.crewMembers": crewMembers,
-        "system.roles": roles,
-      });
-    } else {
-      await this.actor.update({
-        "system.crew.current": crew,
-        "system.crewMembers": crewMembers,
-      });
-      if (this.actor.type == "drone") {
-        ui.notifications.info("Check character sheet for removing drone");
-      }
-    }
-  }
-}
-
-static async _onToggleVehicleStatus(event, target) {
-  event.preventDefault();
-  const dataset = target.dataset;
-  const item = this._getEmbeddedDocument(target);
-  if (item) {
-    console.log(`toggling for  ${item.name}`);
-  }
-  // Handle item rolls.
-  switch (dataset.toggle) {
-    case 'juryRig':
-      console.log("jury");
-      break;
-    case 'broken': 
-      console.log("broken");
-      break;
-    case 'destroyed':
-      console.log('destroyed');
-      break;
-  }
-}
-
-static async _onCrewShow(event, target) {
-  event.preventDefault();
-  const actorId = target.dataset.crewId;
-  const crewActor = game.actors?.get(actorId);
-  crewActor?.sheet?.render(true);
-}
-
-static async _onCrewRoll(event, target) {
-  event.preventDefault();
-  const actorId = target.dataset.crewId;
-  const crewActor = game.actors?.get(actorId);
-  if (!crewActor) {
-    ui.notifications?.error(`Crew no longer exists`);
-    return;
-  }
-  const skills = crewActor.itemTypes.skill;
-  const isChar = crewActor.type == "character" ? true : false;
-  const dialogData = {
-    actor: crewActor,
-    skills: skills,
-    isChar,
-  };
-  const template = "systems/swnr/templates/dialogs/roll-skill-crew.hbs";
-  const html = await renderTemplate(template, dialogData);
-
-  const _rollForm = async (_event, button, html) => {
-    const rollMode = game.settings.get("core", "rollMode");
-    const dice = button.form.elements.dicepool.value;
-    const modifier = parseInt(
-      button.form.elements.modifier?.value
-    ) || 0;
-    const skillId = button.form.elements.skill?.value;
-    const skill = crewActor.getEmbeddedDocument(
-      "Item",
-      skillId
-    );
-    const useNPCSkillBonus =  button.form.elements.useNPCSkillBonus?.checked
-      ? true : false;
-    const npcSkillBonus =
-      useNPCSkillBonus && crewActor.type == "npc"
-        ? crewActor.system.skillBonus
-        : 0;
-    const skillBonus = skill ? skill.system.rank : npcSkillBonus;
-    const statName = button.form.elements.stat
-      ?.value;
-    const stat = crewActor.system["stats"]?.[statName] || {
-      mod: 0,
-    };
-    const formula = `${dice} + @stat + @skillBonus + @modifier`;
-    const roll = new Roll(formula, {
-      skillBonus,
-      modifier,
-      stat: stat.mod,
-    });
-    const skillName = skill ? skill.name : "No Skill";
-    const statNameDisply = statName
-      ? game.i18n.localize("swnr.stat.short." + statName)
-      : "No Stat";
-    const title = `${game.i18n.localize(
-      "swnr.chat.skillCheck"
-    )}: ${statNameDisply}/${skillName}`;
-    await roll.roll({ async: true });
-    roll.toMessage(
-      {
-        speaker: { alias: crewActor.name },
-        flavor: title,
-      },
-      { rollMode }
-    );
-  };
-
-  const d = await foundry.applications.api.DialogV2.prompt({
-    window: { title: game.i18n.format("swnr.dialog.skillRoll", {
-      actorName: crewActor?.name,
-    })},
-    content: html,
-    modal: true,
-    rejectClose: false,
-    ok: {
-      label: game.i18n.localize("swnr.chat.roll"),
-      callback: _rollForm,
-    },
   });
 }
 
@@ -689,7 +891,7 @@ static async _onPay(event, paymentType) {
   this.actor.setScheduledDate(dateObject, paymentType);
 }
 
-async _onHullChange(event) {
+static async _onHullChange(event, target) {
   const targetHull = event.target?.value;
 
   if (targetHull) {
@@ -767,52 +969,49 @@ static async _onRepair(event, target) {
       content: content,
     };
     ChatMessage.create(chatData);
-  } else {
-    ui.notifications.info("Nothing to repair");
   }
 }
 
-static async _onTravel(event, _target) {
+static async _onTravel(event, target) {
   event.preventDefault();
-  if (this.actor.type != "ship") {
-    ui.notifications?.error("Only ships can travel.");
-  };
   if (this.actor.system.spikeDrive.value <= 0) {
     ui.notifications?.error("Drive disabled.");
     return;
   }
   // TODO localize
-  const _doTravel = async (_event, button, _html) => {
-    const days = button.form.elements.amount.value;
-    if (isNaN(parseInt(days))) {
-      ui.notifications?.error(game.i18n.localize("swnr.InvalidNumber"));
-      return;
-    }
-    if (days && days != "") {
-      const nDays = Number(days);
-      if (nDays) {
-        this.actor.system.useDaysOfLifeSupport(nDays);
-        if (game.modules?.get("foundryvtt-simple-calendar")?.active) {
-          this.actor.moveTime(nDays);
+  new Dialog({
+    title: "Travel Days (Use life support)",
+    content: `
+        <form>
+          <div class="form-group">
+            <label>Days of Travel</label>
+            <input type='text' name='inputField'></input>
+          </div>
+        </form>`,
+    buttons: {
+      yes: {
+        icon: "<i class='fas fa-check'></i>",
+        label: `Travel`,
+      },
+    },
+    default: "yes",
+    close: (html) => {
+      const form = html[0].querySelector("form");
+      const days = form.querySelector('[name="inputField"]')?.value;
+      if (days && days != "") {
+        const nDays = Number(days);
+        if (nDays) {
+          this.actor.useDaysOfLifeSupport(nDays);
+          if (game.modules?.get("foundryvtt-simple-calendar")?.active) {
+            this.actor.moveTime(nDays);
+          }
         }
       }
-    }
-  }
-
-  const _proceed = await foundry.applications.api.DialogV2.prompt({
-    window: { title: "Travel Days (Use life support)" },
-    content: `<label>Days of Travel</label><input type='text' name='amount'></input>`,
-    modal: false,
-    rejectClose: false,
-    ok: {
-      callback: _doTravel,
-      icon: 'fas fa-check',
-      label: 'Travel',
-    }
-  });
+    },
+  }).render(true);
 }
 
-static async _onSensor(event, _target) {
+static async _onSensor(event, target) {
   event.preventDefault();
   let defaultCommId = this.actor.system.roles.comms;
   let defaultComm = null;
@@ -855,7 +1054,7 @@ static async _onSensor(event, _target) {
     }
   }
   const dialogData = {
-    actor: this.actor.system,
+    actor: this.actor.data,
     defaultSkill1: "Program",
     defaultSkill2: "Tech/Astronautic",
     defaultStat: "int",
@@ -864,24 +1063,25 @@ static async _onSensor(event, _target) {
     crewArray: crewArray,
   };
 
-  const template = "systems/swnr/templates/dialogs/roll-sensor.hbs";
+  const template = "systems/swnr/templates/dialogs/roll-sensor.html";
   const html = renderTemplate(template, dialogData);
-  const _rollForm = async (_event, button, _html) => {
+  const _rollForm = async (html) => {
+    const form = html[0].querySelector("form");
     const mod = parseInt(
-      button.form.elements.modifier?.value
-    ) || 0;
-    const actorId = button.form.elements.commId.value;
+      form.querySelector('[name="modifier"]')?.value
+    );
+    const actorId = form.querySelector('[name="commId"]')?.value;
     const rollingActor = actorId ? game.actors?.get(actorId) : null;
-    const dice = button.form.elements.dicepool.value;
-    const skillName = button.form.elements.skill.value;
-    const statName = button.form.elements.stat.value;
+    const dice = form.querySelector('[name="dicepool"]').value;
+    const skillName = form.querySelector('[name="skill"]')?.value;
+    const statName = form.querySelector('[name="stat"]')?.value;
     const targetModifier = parseInt(
-      button.form.elements.targetModifier?.value
-    ) || 0;
+      form.querySelector('[name="targetModifier"]')?.value
+    );
     const observerModifier = parseInt(
-      button.form.elements.observerModifier?.value
-    ) || 0;
-    const rollingAs = button.form.elements.rollingAs.value;
+      form.querySelector('[name="observerModifier"]')?.value
+    );
+    const rollingAs = form.querySelector('[name="rollingAs"]')?.value;
     if (
       rollingAs != "observer" &&
       rollingAs != "target" &&
@@ -890,7 +1090,7 @@ static async _onSensor(event, _target) {
       ui.notifications?.error("Error with rolling as ");
       return;
     }
-    const rollType = button.form.elements.rollType.value;
+    const rollType = form.querySelector('[name="rollType"]')?.value;
     if (
       rollType != "roll" &&
       rollType != "gmroll" &&
@@ -906,9 +1106,8 @@ static async _onSensor(event, _target) {
     if (rollingActor) {
       if (skillName) {
         // We need to look up by name
-        console.log("looking for skill", skillName); //TODO remove
         for (const skill of rollingActor.itemTypes.skill) {
-          if (skillName == skill.name) {
+          if (skillName == skill.data.name) {
             skillMod =
               skill.system["rank"] < 0 ? -1 : skill.system["rank"];
           }
@@ -923,7 +1122,7 @@ static async _onSensor(event, _target) {
       }
       actorName = rollingActor.name;
     }
-    this.actor.system.rollSensor(
+    this.actor.rollSensor(
       actorName,
       targetModifier,
       observerModifier,
@@ -935,19 +1134,24 @@ static async _onSensor(event, _target) {
     );
   };
 
-
-  const _proceed = await foundry.applications.api.DialogV2.prompt({
-      window: { title: title },
+  this.popUpDialog?.close();
+  this.popUpDialog = new ValidatedDialog(
+    {
+      title: title,
       content: await html,
-      modal: false,
-      rejectClose: false,
-      ok: {
+      default: "roll",
+      buttons: {
+        roll: {
           label: game.i18n.localize("swnr.chat.roll"),
           callback: _rollForm,
+        },
       },
+    },
+    {
+      classes: ["swnr"],
     }
   );
-  
+  this.popUpDialog.render(true);
 }
 
 static async _onSpike(event, target) {
@@ -1003,7 +1207,7 @@ static async _onSpike(event, target) {
   }
 
   const dialogData = {
-    actor: this.actor.system,
+    actor: this.actor.data,
     defaultSkill1: "Pilot",
     defaultSkill2: "Navigation",
     defaultStat: "int",
@@ -1013,24 +1217,25 @@ static async _onSpike(event, target) {
     baseDifficulty: 7,
   };
 
-  const template = "systems/swnr/templates/dialogs/roll-spike.hbs";
+  const template = "systems/swnr/templates/dialogs/roll-spike.html";
   const html = renderTemplate(template, dialogData);
 
-  const _rollForm = async (_event, button, _html) => {
+  const _rollForm = async (html) => {
+    const form = html[0].querySelector("form");
     const mod = parseInt(
-      button.form.elements.modifier?.value
-    ) || 0;
-    const pilotId = button.form.elements.pilotId.value;
+      form.querySelector('[name="modifier"]')?.value
+    );
+    const pilotId = form.querySelector('[name="pilotId"]')?.value;
     const pilot = pilotId ? game.actors?.get(pilotId) : null;
-    const dice = button.form.elements.dicepool.value;
-    const skillName = button.form.elements.skill.value;
-    const statName = button.form.elements.stat.value;
+    const dice = form.querySelector('[name="dicepool"]').value;
+    const skillName = form.querySelector('[name="skill"]')?.value;
+    const statName = form.querySelector('[name="stat"]')?.value;
     const difficulty = parseInt(
-      button.form.elements.difficulty?.value
-    ) || 0;
+      form.querySelector('[name="difficulty"]')?.value
+    );
     const travelDays = parseInt(
-      button.form.elements.travelDays?.value
-    ) || 0;
+      form.querySelector('[name="travelDays"]')?.value
+    );
     let skillMod = 0;
     let statMod = 0;
     let pilotName = "";
@@ -1038,7 +1243,7 @@ static async _onSpike(event, target) {
       if (skillName) {
         // We need to look up by name
         for (const skill of pilot.itemTypes.skill) {
-          if (skillName == skill.name) {
+          if (skillName == skill.data.name) {
             skillMod =
               skill.system["rank"] < 0 ? -1 : skill.system["rank"];
           }
@@ -1053,7 +1258,7 @@ static async _onSpike(event, target) {
       }
       pilotName = pilot.name;
     }
-    this.actor.system.rollSpike(
+    this.actor.rollSpike(
       pilotId,
       pilotName,
       skillMod,
@@ -1065,19 +1270,24 @@ static async _onSpike(event, target) {
     );
   };
 
-  const popUpDialog = await foundry.applications.api.DialogV2.prompt(
+  this.popUpDialog?.close();
+  this.popUpDialog = new ValidatedDialog(
     {
-      window: { title: title },
+      title: title,
       content: await html,
-      modal: false,
-      rejectClose: false,
-      ok: {
+      default: "roll",
+      buttons: {
+        roll: {
           label: game.i18n.localize("swnr.chat.roll"),
           callback: _rollForm,
+        },
       },
+    },
+    {
+      classes: ["swnr"],
     }
   );
-  
+  this.popUpDialog.render(true);
 }
 
 static async _onRefuel(event, target) {
@@ -1103,7 +1313,7 @@ static async _onRefuel(event, target) {
 
 static _onCrisis(event, target) {
   event.preventDefault();
-  this.actor.system.rollCrisis();
+  this.actor.rollCrisis(); //TODO
 }
 
 static async _onSysFailure(event, target) {
@@ -1112,23 +1322,24 @@ static async _onSysFailure(event, target) {
     actorName: this.actor?.name,
   });
   const dialogData = {};
-  const template = "systems/swnr/templates/dialogs/roll-ship-failure.hbs";
+  const template = "systems/swnr/templates/dialogs/roll-ship-failure.html";
   const html = renderTemplate(template, dialogData);
 
-  const _rollForm = async (_event, button, html) => {
-    const incDrive = button.form.elements.incdrive?.checked
+  const _rollForm = async (html) => {
+    const form = html[0].querySelector("form");
+    const incDrive = form.querySelector('[name="inc-drive"]')?.checked
       ? true
       : false;
-    const incWpn = button.form.elements.incwpn?.checked
+    const incWpn = form.querySelector('[name="inc-wpn"]')?.checked
       ? true
       : false;
-    const incFit = button.form.elements.incfit?.checked
+    const incFit = form.querySelector('[name="inc-fit"]')?.checked
       ? true
       : false;
-    const incDef = button.form.elements.incdef?.checked
+    const incDef = form.querySelector('[name="inc-def"]')?.checked
       ? true
       : false;
-    const whatToRoll = button.form.elements.what.value;
+    const whatToRoll = form.querySelector('[name="what"]')?.value;
 
     const sysToInclude = [];
     if (incDrive) {
@@ -1143,21 +1354,27 @@ static async _onSysFailure(event, target) {
     if (incDef) {
       sysToInclude.push("def");
     }
-    this.actor.system.rollSystemFailure(sysToInclude, whatToRoll);
+    this.actor.rollSystemFailure(sysToInclude, whatToRoll);
   };
 
-  const popUpDialog = await foundry.applications.api.DialogV2.prompt(
+  this.popUpDialog?.close();
+  this.popUpDialog = new ValidatedDialog(
     {
-      window: { title: title },
+      title: title,
       content: await html,
-      modal: false,
-      rejectClose: false,
-      ok: {
+      default: "roll",
+      buttons: {
+        roll: {
           label: game.i18n.localize("swnr.chat.roll"),
           callback: _rollForm,
+        },
       },
+    },
+    {
+      classes: ["swnr"],
     }
   );
+  this.popUpDialog.render(true);
 }
 
 static _onCalcCost(event, target) {
@@ -1232,25 +1449,13 @@ static async _setCaptSupport(dept) {
   });
 }
 
-  /* Handles drag-and-drop visual */
-  _handleDragEnter(event){
-    let ele = event.target.closest(".item-row");
-    if(ele) $(ele).addClass('over')
-  }
-  
-  _handleDragLeave(event) {
-    let ele = event.target.closest(".item-row");
-    if(ele) $(ele).removeClass('over')
-  }
-
-
-async _onShipAction(event) {
+static async _onShipAction(event, target) {
   event.preventDefault();
   const actionName = event.target?.value;
   if (actionName === "") {
     return;
   }
-  const action = CONFIG.SWN.shipActions[actionName];
+  const action = ACTIONS[actionName];
   if (!action) {
     ui.notifications?.error("There was an error in looking up your action");
     return;
@@ -1270,23 +1475,15 @@ async _onShipAction(event) {
   if (actionName == "startRound") {
     let depts = ``;
     let roleOrder = [];
-    let roles = [];
-    if (this.actor.system.roleOrder && this.actor.system.roleOrder.length > 0) {
+    if (this.actor.system.roleOrder) {
       // we may have set an order prior
       roleOrder = this.actor.system.roleOrder;
     } else {
       // get the default list
-      // to iterate schema      
-      // for (const role of Object.keys(this.actor.system.schema.fields.roles.fields)) {
-      for (const role of Object.entries(this.actor.system.roles)) {
-        roleOrder.push(role[0]);
+      for (const role in this.actor.system.roles) {
+        roleOrder.push(role);
       }
     }
-    if (roleOrder.length == 0) {
-      ui.notifications?.error("No roles assigned");
-      return;
-    }
-    let counter = 0;
     for (const role of roleOrder) {
       let roleName = "";
       if (this.actor.system.roles[role]) {
@@ -1296,77 +1493,47 @@ async _onShipAction(event) {
         }
       }
       depts += `<div class="border p-2 flex border-black role-order" data-role="${role}" data-role-name="${roleName}"><a><i class="fas fa-sort"></i></a>${role}${roleName}</div>`;
-      let capRole = role.charAt(0).toUpperCase() + role.slice(1).toLowerCase();
-      let roleTuple = [role, `${capRole}${roleName}`, counter++];
-      roles.push(roleTuple);
     }
-    let selectDialog = `
-    <div class="flex flex-col">
-      <h2> Order Departments/Roles </h2>
-      <input type="hidden" id="roleCount" value="${roles.length}">
-    `;
-    for (let i = 1; i <= roles.length; i++) {
-      selectDialog += `<div class="flex flexrow"> ${i}: <select id="role${i}">`;
-      for (let j = 0; j < roles.length; j++) {
-        let role = roles[j];
-        let selected = '';
-        if (role[2] == i - 1) {
-          selected = 'selected';
-        }
-        selectDialog += `<option value="${role[0]}" ${selected}>${role[1]}</option>`;
-      }
-      selectDialog += `</select></div>`;
-    }
-    selectDialog += `</div>`;
-    const sortedDialogTemplate = `
-    <div class="flex flex-col">
-      <h2> Order Departments/Roles </h2>
+    const dialogTemplate = `
+    <div class="flex flex-col -m-2 p-2 pb-4 space-y-2">
+      <h1> Order Departments/Roles </h1>
       <div class="flex flexrow">
           <div id="deptOrder">
           ${depts}
           </div>
       </div>
       <script>
-      console.log("hello");
-
       var el = document.getElementById('deptOrder');
-      console.log(el);
-      el.
-      el.find('.role-order').each((i, di) {
-        console.log(di);
-        new Dragster( di );
-        li.addEventListener("dragster:enter", ev => this._handleDragEnter(ev) , false);
-        li.addEventListener("dragster:leave", ev => this._handleDragLeave(ev) , false);
-      });
-
-      // var sortable = Sortable.create(el);
+      var sortable = Sortable.create(el);
       </script>
     </div>
     `;
-    const d = await foundry.applications.api.DialogV2.prompt({
-      window: { title: "Set Order"},
-      content: selectDialog,
-      modal: true,
-      rejectClose: false,
-      ok: {
-        label: "Set Order",
-        callback: async (_event, button, html) => {
-          let count = button.form.elements.roleCount.value;
-          const orderArr = [];
-          for (let i = 1; i <= count; i++) {
-            const role = button.form.elements[`role${i}`].value;
-            if (orderArr.indexOf(role) >= 0) {
-              ui.notifications?.error("Duplicate role found. Not setting order.");
-              return;
-            }
-            orderArr.push(role);
-          }
-          if (orderArr.length > 0) {
-            await this.actor.update({ system: { roleOrder: orderArr } });
-          }
-        }
-      }
-    });
+    new Dialog(
+      {
+        title: "Set Order",
+        content: dialogTemplate,
+        buttons: {
+          setOrder: {
+            label: "Set Order",
+            callback: async (html) => {
+              const order = html.find("#deptOrder");
+              const orderArr = [];
+              order.children(".role-order").each(function () {
+                orderArr.push($(this).data("role"));
+              });
+              if (orderArr.length > 0) {
+                await this.actor.update({ data: { roleOrder: orderArr } });
+              }
+            },
+          },
+          close: {
+            label: "Close",
+          },
+        },
+        default: "setOrder",
+      },
+      { classes: ["swnr"] }
+    ).render(true);
     return;
   } else if (actionName == "endRound") {
     // endRound action is special. clear and reset.
@@ -1379,7 +1546,7 @@ async _onShipAction(event) {
     if (actionsTaken.length > 0) {
       for (const act of actionsTaken) {
         const actTitle =
-          CONFIG.SWN.shipActions[act] && CONFIG.SWN.shipActions[act].title ? CONFIG.SWN.shipActions[act].title : act;
+          ACTIONS[act] && ACTIONS[act].title ? ACTIONS[act].title : act;
         actionsText += `<li>${actTitle}</li>`;
       }
       actionsText += "</ul>";
@@ -1439,7 +1606,7 @@ async _onShipAction(event) {
           foundActor = true;
           if (defaultActor.type == "character") {
             for (const skill of defaultActor.itemTypes.skill) {
-              if (action.skill == skill.name) {
+              if (action.skill == skill.data.name) {
                 skillLevel = skill.system["rank"];
                 dicePool =
                   skill.system["pool"] && skill.system["pool"] != "ask"
@@ -1560,4 +1727,41 @@ async _onShipAction(event) {
   return;
 }
 
+
+
+
+
+  /********************
+   *
+   * Actor Override Handling
+   *
+   ********************/
+
+  /**
+   * Submit a document update based on the processed form data.
+   * @param {SubmitEvent} event                   The originating form submission event
+   * @param {HTMLFormElement} form                The form element that was submitted
+   * @param {object} submitData                   Processed and validated form data to be used for a document update
+   * @returns {Promise<void>}
+   * @protected
+   * @override
+   */
+  async _processSubmitData(event, form, submitData) {
+    const overrides = foundry.utils.flattenObject(this.actor.overrides);
+    for (let k of Object.keys(overrides)) delete submitData[k];
+    await this.document.update(submitData);
+  }
+
+  /**
+   * Disables inputs subject to active effects
+   */
+  #disableOverrides() {
+    const flatOverrides = foundry.utils.flattenObject(this.actor.overrides);
+    for (const override of Object.keys(flatOverrides)) {
+      const input = this.element.querySelector(`[name="${override}"]`);
+      if (input) {
+        input.disabled = true;
+      }
+    }
+  }
 }
