+++ src/scss/components/_items.scss
@@ -23,10 +23,6 @@
     flex: 1 1 auto;
     // font-size: 16px;
   }
-  .item-name-long {
-    flex: 2 1 100px;
-
-  }
 }
 
 // -----------------------------------------
@@ -64,12 +60,6 @@
     }
   }
 
-
-  // Item Name
-  .item-name-long {
-    flex: 2 1 100px;
-  }
-
   // Control Buttons
   .item-controls {
     display: flex;
@@ -84,7 +74,7 @@
 
   .item-controls-small {
     display: flex;
-    flex: 0 0 50px;
+    flex: 0 0 20px;
     justify-content: flex-end;
     a {
       font-size: 12px;
@@ -168,13 +158,8 @@
   padding: 0 8px;
 }
 
-.item-type {
-  flex: 0 0 60px;
-  padding: 0 8px;
-}
-
 .item-controls-small {
-  flex: 0 1 50px;
+  flex: 0 1 40px;
   padding: 0 8px;
 }
 .item-loc {
