--- templates/actor/cyberdeck/cyberdeckHeader.hbs
@@ -1,114 +1,0 @@
-{{! Sheet Header }}
-<header class='sheet-header'>
-  <div class="profile-img"><!-- Profile Image -->
-    <img
-      class='profile-img'
-      src='{{actor.img}}'
-      data-edit='img'
-      data-action='onEditImage'
-      title='{{actor.name}}'
-      height='100'
-      width='100'
-    />
-
-    <div class="resource-bar"><!-- The character health bar -->
-      <div style="width:{{system.hackerHP.percentage}}%" class="health-fill"></div>
-        <div class="text">
-          HP
-            <input
-              type='text'
-              name='system.health.value'
-              value='{{system.hackerHP.value}}'
-              data-dtype='Number'
-              class='nested-field short-field align-right current-value'
-            />
-          <p>/</p>
-            <input
-              type='text'
-              name='system.health.max'
-              value='{{system.hackerHP.max}}'
-              data-dtype='Number'
-              class='nested-field short-field'
-            />
-        </div>
-    </div>
-  </div>
-  <div class='header-fields'>
-    <div class='document-name'>
-      <input
-        name='name'
-        type='text'
-        value='{{actor.name}}'
-        placeholder='Name'
-      />
-    </div>
-    <div class='resources grid grid-3col'>
-        <div
-          class='resource grid-span-3 thin-bar'
-          style="text-align:right;"
-        >
-          <label class='resource-label'>{{localize 'swnr.item.favorite'}}</label>
-          <input
-            type="checkbox"
-            class="small-checkbox"
-            name="system.favorite"
-            {{#if system.favorite}}checked{{/if}}
-          />
-        </div>
-      {{! "flex-group-center" is also defined in the _grid.scss file
-        and it will add a small amount of padding, a border, and will
-        center all of its child elements content and text. }}
-    <div class='resources grid grid-4col'>
-      <div>
-        {{formGroup systemFields.health value=system.health localize=true widget=headerWidget}}
-        <a class="item-control" data-action="refreshShielding" title='{{localize 'swnr.sheet.cyberdeck.refreshShielding'}}'><i class="fas fa-hourglass-end"></i></a>
-      </div>
-      {{formGroup systemFields.memory value=system.memory localize=true widget=headerWidget}}
-      {{formGroup systemFields.cpu value=system.cpu localize=true widget=headerWidget}}
-      <div>
-        {{formGroup systemFields.access value=access localize=true widget=headerWidget}}
-        <a class="item-control" data-action="refreshAccess" title='{{localize 'swnr.sheet.cyberdeck.refreshAccess'}}'><i class="fas fa-hourglass-end"></i></a>
-      </div>
-    </div>
-
-    <div class="resource grid grid-3col thin-bar">
-      <div class='resource' style="">
-        <label class='resource-label'>
-          {{localize 'SWN.Actor.Cyberdeck.FIELDS.neuralBuffer.label'}}
-          </label>
-        <input type="checkbox" class="small-checkbox" name="system.neuralBuffer" {{#if system.neuralBuffer}}checked{{/if}} />
-      </div>
-      <div class='resource' style="">
-        <label class='resource-label'>
-          {{localize 'SWN.Actor.Cyberdeck.FIELDS.wirelessConnectionPenalty.label'}}
-          </label>
-        <input type="checkbox" class="small-checkbox" name="system.wirelessConnectionPenalty" {{#if system.wirelessConnectionPenalty}}checked{{/if}} />
-      </div>
-      <div class='resource' style="">
-        <label class='resource-label'>
-          {{localize 'SWN.Actor.Cyberdeck.FIELDS.crownPenalty.label'}}
-          </label>
-        <input type="checkbox" class="small-checkbox" name="system.crownPenalty" {{#if system.crownPenalty}}checked{{/if}} />
-      </div>
-    </div>
-
-    <div class="resource grid grid-4col thin-bar">
-      {{formGroup systemFields.baseShielding value=system.baseShielding localize=true widget=headerWidget}}
-      {{formGroup systemFields.bonusShielding value=system.bonusShielding localize=true widget=headerWidget}}
-      {{formGroup systemFields.bonusAccess value=system.bonusAccess localize=true widget=headerWidget}}
-    </div>
-    <div class="resource grid grid-4col thin-bar">
-      {{formGroup systemFields.cost value=system.cost localize=true widget=headerWidget}}
-      {{formGroup systemFields.encumberance value=system.encumberance localize=true widget=headerWidget}}
-      {{formGroup systemFields.hacker value=hacker.name localize=true widget=headerWidget}}
-      <div class="item-controls">
-        <a class="item-control" data-action="hackerRoll" title='{{localize ' swnr.sheet.roll-skill'}}'><i class="fas fa-dice"></i></a>
-        <a class="item-control" data-action="hackerDelete" title='{{localize 'swnr.sheet.delete-item'}}'><i class="fas fa-trash"></i></a>
-      </div>
-    </div>
-
-
-         
-    </div>
-  </div>
-</header>
