--- templates/actor/cyberdeck/programs-list.hbs
@@ -1,139 +1,0 @@
-    <ol class='programs-list'>
-      <li class='item flex flexrow items-header'>
-        <div class="item-name"> 
-          {{#if (eq subType "verb")}}
-            {{ localize 'swnr.sheet.program.verb' }}
-          {{else if (eq subType "subject")}}
-            {{ localize 'swnr.sheet.program.subject' }}
-          {{else if (eq subType "running")}}
-            {{ localize 'swnr.sheet.program.running' }}
-          {{else if (eq subType "dataFile")}}
-            {{ localize 'swnr.sheet.program.dataFile' }}
-          {{/if}}
-        </div>
-
-        {{#if (eq subType "running")}}
-          <div class="item-name">
-            {{ localize 'swnr.sheet.program.use' }}
-          </div>
-        {{/if}}
-
-        {{#if (eq subType "verb")}}
-          <div class="item-name">
-            {{ localize 'swnr.sheet.program.use' }}
-          </div>
-          <div class="item-name">
-            {{ localize 'swnr.sheet.program.targetAllowedShort' }}
-          </div>
-          <div class="item-name">
-            {{ localize 'swnr.sheet.program.accessCostShort' }}
-          </div>
-        {{/if}}
-
-        {{#if (eq subType "subject")}}
-          <div class="item-name">
-            {{ localize 'swnr.sheet.program.subjectType' }}
-          </div>
-        {{/if}}
-
-
-        <div class='item-controls'>
-          {{#if @root.editable}}
-            <a 
-              class="item-control item-search"
-              title="{{localize 'swnr.sheet.search-item'}}"
-              data-item-type="{{ type }}" 
-              data-item-name="{{ name }}">
-            <i class="fas fa-search"></i>
-            </a>
-            {{#if (eq subType "running")}}
-              <a
-                class='item-control item-create'
-                title='{{localize "DOCUMENT.Create" type='Item'}}'
-                data-action='runProgram'
-                data-document-class='Item'
-                data-type='program'
-                data-system.type='{{subType}}'
-              >
-                <i class='fas fa-plus'></i>
-                {{localize 'swnr.sheet.program.run'}}
-              </a>
-            {{else}}
-              <a
-                class='item-control item-create'
-                title='{{localize "DOCUMENT.Create" type='Item'}}'
-                data-action='createDoc'
-                data-document-class='Item'
-                data-type='program'
-                data-system.type='{{subType}}'
-              >
-                <i class='fas fa-plus'></i>
-                {{localize 'DOCUMENT.New' type=type}}
-              </a>
-            {{/if}}
-          {{/if}}
-        </div>
-      </li>
-      {{#each itemList as |item id|}}
-        {{#if (eq ../subType item.system.type)}}
-            <li
-              class='item flex flexrow'
-              data-item-id='{{item._id}}'
-              data-drag='true'
-              data-document-class='Item'
-            >
-              <div class='item-name'>
-                <div class='item-image'>
-                  <a class='rollable' data-roll-type='item' data-action='roll'>
-                    <img
-                      src='{{item.img}}'
-                      title='{{item.name}}'
-                      width='24'
-                      height='24'
-                    />
-                  </a>
-                </div>
-                <div>{{item.name}}
-                </div>
-              </div>
-            {{#if (eq system.type "running")}}
-              <div class="item-name">
-                {{ item.system.useAffects }}
-              </div>
-            {{else if (eq system.type "verb")}}
-              <div class="item-name">
-                {{ item.system.useAffects }}
-              </div>
-              <div class="item-name">
-                {{ item.system.target }}
-              </div>
-              <div class="item-name">
-                {{ item.system.accessCost }}
-              </div>
-            {{else if (eq system.type "subject")}}
-              <div class="item-name">
-                {{ item.system.target }}
-              </div>
-            {{/if}}
-              <div class='item-controls'>
-                <a
-                  class='item-control item-edit'
-                  title='{{localize "DOCUMENT.Update" type='Item'}}'
-                  data-action='viewDoc'
-                >
-                  <i class='fas fa-edit'></i>
-                </a>
-                {{#if @root.editable}}
-                  <a
-                    class='item-control item-delete'
-                    title='{{localize "DOCUMENT.Delete" type='Item'}}'
-                    data-action='deleteDoc'
-                  >
-                    <i class='fas fa-trash'></i>
-                  </a>
-                {{/if}}
-              </div>
-            </li>
-        {{/if}}
-      {{/each}}
-    </ol>
