--- templates/actor/fragments/crew-list.hbs
@@ -1,22 +1,0 @@
-<div>
-  <li class="item-titles flexrow">
-    <!--<div class="item-caret"><i class="fas fa-caret-down"></i></div>-->
-    <div class="item-name">{{localize 'swnr.sheet.ship.crew-name'}} </div>
-    <div class="item-controls">
-    </div>
-  </li>
-  <ol class="item-list">
-    {{#each crewArray as | crew id| }}
-    <li class="item-entry">
-      <div class="item flexrow" data-crew-id="{{crew.id}}" data-crew-name="{{crew.name}}">
-        <div class="item-name flexrow" data-action="crewShow" data-crew-id="{{crew.id}}">
-          <a class="crew-show">{{crew.name}}</a>
-        </div>
-        <div class="item-controls">
-          <a class="item-control" data-action="crewRoll" data-crew-id="{{crew.id}}" title='{{localize ' swnr.sheet.roll-skill'}}'><i class="fas fa-dice"></i></a>
-          <a class="item-control" data-action="crewDelete" data-crew-id="{{crew.id}}" title='{{localize 'swnr.sheet.delete-item'}}'><i class="fas fa-trash"></i></a>
-        </div>
-    </li>
-    {{/each}}
-  </ol>
-</div>
