--- templates/actor/fragments/vehicle-weapon-table.hbs
@@ -1,68 +1,0 @@
-<!-- weapons -->
-<div>
-  <ol class="items-list">
-    <li class='flex flexrow items-header'>
-      <!--<div class="item-caret"><i class="fas fa-caret-down"></i></div>-->
-      <div class="item-name">{{localize 'swnr.sheet.ship.weapon'}}</div>
-      <div class="item-dmg">{{localize 'swnr.sheet.ship.dmg'}}</div>
-      <div class="item-qualities">{{localize 'swnr.sheet.ship.qualities'}}</div>
-      <div class="item-controls text-right">
-        <a class="item-control item-search" title="{{localize 'swnr.sheet.search-item'}}" data-item-type="shipWeapon"
-          data-item-name="Weapon" data-item-subtype={{subtype}}>
-          <i class="fas fa-search"></i>
-        </a>
-        <a class="item-control item-create" title="{{localize 'swnr.sheet.add-item'}}" data-action='createDoc'
-          data-document-class='Item' data-type="shipWeapon" data-item-name="Weapon">
-          <i class="fas fa-plus"></i>
-        </a>
-      </div>
-    </li>
-    {{#each actor.itemTypes.shipWeapon as |item id|}}
-    <li class='item flex flexrow' data-item-id='{{item._id}}' data-drag='true' data-document-class='Item'>
-      <div class="item-name item-click item-rollable flexrow flex2">
-        <div class="item-image" style="background-image: url({{item.img}})"></div>
-          <a class='rollable' data-roll-type='item' data-action='roll'>
-            {{item.name}}
-          </a>
-      </div>
-      <div>
-        {{item.system.damage}}
-      </div>
-      <div>
-        {{item.system.qualities}}
-      </div>
-      <div class="item-controls  text-right">
-        {{#if (ne item.system.ammo.type "none")}}
-        <a class="item-control item-reload " title="{{localize 'swnr.sheet.reload-item'}}">
-          <i class="fas fa-retweet"></i>
-        </a>
-        {{/if}}
-        <a class="item-control item-toggle-jury  {{#unless item.system.juryRigged}}item-working{{/unless}}"
-          title='{{localize 'swnr.sheet.jury-rig'}}' data-action='toggle' data-toggle="juryRig">
-          <i class="fas fa-tape {{#unless item.system.juryRigged}}item-working{{/unless}} "></i>
-        </a>
-        <a class="item-control item-toggle-broken  {{#unless item.system.broken}}item-working{{/unless}}"
-          title='{{localize ' swnr.sheet.break-item'}}' data-action='toggle' data-toggle="broken">
-          <i class="fas fa-heart-broken {{#unless item.system.broken}}item-working{{/unless}}"></i>
-        </a>
-        <a class="item-control item-toggle-destroy  {{#unless item.system.destroyed}}item-working{{/unless}}"
-          title='{{localize ' swnr.sheet.destroy-item'}}' data-action='toggle' data-toggle="destroyed">
-          <i class="fas fa-skull-crossbones {{#unless item.system.destroyed}}item-working{{/unless}}"></i>
-        </a>
-        <div class='item-controls'>
-          {{#if @root.editable}}
-
-          <a class='item-control item-edit' title='{{localize "DOCUMENT.Update" type=' Item'}}' data-action='viewDoc'>
-            <i class='fas fa-edit'></i>
-          </a>
-          <a class='item-control item-delete' title='{{localize "DOCUMENT.Delete" type=' Item'}}'
-            data-action='deleteDoc'>
-            <i class='fas fa-trash'></i>
-          </a>
-          {{/if}}
-      </div>
-</div>
-</li>
-{{/each}}
-</ol>
-</div>
