+++ templates/actor/vehicle/drone.hbs
@@ -17,10 +17,6 @@
       <div class="resource">
         {{formGroup systemFields.moveType value=system.moveType localize=true}}
       </div>
-
-      <div class="resource">
-        {{formGroup systemFields.tl value=system.tl localize=true}}
-      </div>
     </div><!-- end grid-col SWNRDroneData extends SWNRVehicleTemplateBase  -->
   </div>
 </section>
