+++ templates/actor/vehicle/mech.hbs
@@ -11,10 +11,6 @@
   <div class="resource">
    {{formGroup systemFields.mechHullType value=system.mechHullType localize=true}}
   </div>
-
-  <div class="resource">
-    {{formGroup systemFields.tl value=system.tl localize=true}}
-  </div>
 </div><!-- end grid-col SWNRMechData extends SWNRVehicleTemplateBase -->
 
   </div>
