+++ templates/actor/vehicle/ship.hbs
@@ -52,12 +52,14 @@
         {{!-- <div class="resource">
         {{formGroup systemFields.cargoCarried value=system.cargoCarried localize=true}}
         </div> --}}
-
-        {{!-- <div class="resource">
-        {{formGroup systemFields.crewSkillBonus value=system.crewSkillBonus localize=true}}
-        </div> --}}
         <div class="resource">
-        {{formGroup systemFields.tl value=system.tl localize=true}}
+        {{formGroup systemFields.commandPoints value=system.commandPoints localize=true}}
+        </div>
+        <div class="resource">
+        {{formGroup systemFields.npcCommandPoints value=system.npcCommandPoints localize=true}}
+        </div>
+        <div class="resource">
+        {{formGroup systemFields.crewSkillBonus value=system.crewSkillBonus localize=true}}
         </div>
         {{!-- <div class="resource">
         {{formGroup systemFields.actionsTaken value=system.actionsTaken localize=true}}
