+++ templates/actor/vehicle/vehicle.hbs
@@ -7,10 +7,6 @@
       <div class="resource">
         {{formGroup systemFields.tonnage value=system.tonnage localize=true}}
       </div>
-
-      <div class="resource">
-      {{formGroup systemFields.tl value=system.tl localize=true}}
-      </div>
       {{!-- <div class="resource">
         {{formGroup systemFields.size value=system.size localize=true}}
       </div> --}}
