--- templates/chat/sensor-roll.hbs
@@ -1,8 +1,0 @@
-<div>
-  <h3>Sensor Roll as {{ typeDesc }}</h3>
-  <span title="{{ skillRollStr }}">{{{ skillRoll }}}</span>
-  {{#if opposedRoll}}
-  <span> Other Ship ({{ otherDesc }}) </span>
-  {{{opposedRoll}}}
-  {{/if}}
-</div>
