--- templates/chat/spike-roll.hbs
@@ -1,12 +1,0 @@
-<div>
-  <h3>Spike Drill</h3>
-  <span title="{{ skillRollStr }}">{{{ skillRoll }}}</span>
-  {{#if pass}}
-    <span><b>Spike drill successful.</b></span><br>
-  {{else}}
-    <span><b>Spike drill failed!</b></span><br>
-   <span>  {{{ failRoll}}}</span>
-   {{ failText}}<br>
-  {{/if}}
-   <i>Travel Days: {{ travelDays}}</i>
-</div>
