--- templates/dialogs/add-currency.hbs
@@ -1,8 +1,0 @@
-<form class="currency-add-dialog">
-  <div class="form-group">
-    <label for="currency-add-dialog-amount">Amount to Add/Subtract:</label>
-    <input type="number" name="amount" 
-      title="Insert a value to be added to the currency type. Negative values will be deducted." 
-        autofocus />
-  </div>
-</form>
