--- templates/dialogs/roll-ship-attack.hbs
@@ -1,80 +1,0 @@
-<form class="flex">
-  <div class="flex flex-col  flex-grow">
-    <div class="flex flexrow gap-2 flex-grow">
-
-      <div class="flex flex-col">
-        <label for="shooterId">{{localize 'swnr.sheet.ship.shooter'}}:</label>
-        <select name="shooterId">
-          {{#select gunnerId}}
-          {{#each crewArray }}
-          <option value="{{this.id}}">{{this.name}}</option>
-          {{/each}}
-          {{/select}}
-        </select>
-      </div>
-
-
-
-      <div class="flex flex-col">
-        <label for="skill">{{localize 'swnr.sheet.skill'}}: </label>
-        <select name="skill">
-          <option value="">{{ localize 'swnr.weapon.noSetSkill'}}</option>
-          {{#each gunner.itemTypes.skill }}
-          <option value="{{this.name}}" 
-            {{#if (eq this.name ../defaultSkill1)}} selected {{/if}}
-            {{#if (eq this.name ../defaultSkill2)}} selected {{/if}}
-          >{{this.name}}</option>
-          {{/each}}
-        </select>
-      </div>
-
-
-      {{!-- <div class="flex flex-col">
-        {{! dice }}
-        <label for="">{{localize 'swnr.chat.dicepool'}}: </label>
-        <select name="dicepool">
-          <option value="2d6">2D6</option>
-          <option value="3d6kh2">3D6</option>
-          <option value="4d6kh2">4D6</option>
-        </select>
-      </div> --}}
-
-      <div class="flex flex-col">
-        {{! stat}}
-        <label for="stat">{{localize 'swnr.sheet.stats'}}: </label>
-        <select name="stat">
-          {{#select defaultStat}}
-          {{#each gunner.system.stats as |box name |}}
-          <option value="{{name}}">{{localize (concat "swnr.stat.long." name)}}</option>
-          {{/each}}
-          {{/select}}
-        </select>
-      </div>
-
-    </div>
-    <div class="flex flexrow gap-2  flex-grow">
-
-      <div class="flex flex-col">
-        <label for="weaponAb">{{localize 'swnr.sheet.attack-bonus'}}: </label>
-        <input name="weaponAb" type="number" step="1" value="{{ weapon.ab }}"/>
-      </div>
-
-      <div class="flex flex-col">
-        <label for="npcSkill">{{localize 'swnr.sheet.ship.crew-skill-bonus'}}: </label>
-        <input name="npcSkill" type="number" step="1" value="{{ actor.data.crewSkillBonus }}"/>
-      </div>
-
-      <div class="flex flex-col">
-        <label for="modifier">{{localize 'swnr.chat.modifier'}}: </label>
-        <input name="modifier" type="number" step="1" value="0"/>
-      </div>
-
-      {{#if (and weapon.ammo.burst (ne 'none' weapon.ammo.type)) }}
-      <div class="flex flex-col ">
-        <label for="burstFire">{{localize 'swnr.weapon.burst'}}: </label>
-        <input type="checkbox" name="burstFire" id="burstFire" {{checked burstFireHasAmmo}}>
-      </div>
-      {{/if}}
-    </div>
-  </div>
-</form>
