--- templates/dialogs/roll-ship-failure.hbs
@@ -1,33 +1,0 @@
-<form class="flex">
-  <div class="flex flex-col">
-    <div class="flex flexrow">
-        <label for="incdrive">{{localize 'swnr.dialog.failure.inc-drive'}}:</label>
-        <input type="checkbox" name="incdrive" id="incdrive" checked />
-    </div>
-    <div class="flex flexrow">
-        <label for="incwpn">{{localize 'swnr.dialog.failure.inc-wpn'}}:</label>
-        <input type="checkbox" name="incwpn" id="incwpn" checked />
-    </div>
-    <div class="flex flexrow">
-        <label for="incfit">{{localize 'swnr.dialog.failure.inc-fit'}}:</label>
-        <input type="checkbox" name="incfit" id="incfit" checked />
-    
-
-    </div>
-    <div class="flex flexrow">
-    <label for="incdef">{{localize 'swnr.dialog.failure.inc-def'}}:</label>
-        <input type="checkbox" name="incdef" id="incdef" />
-    
-
-    </div>
-    <div class="flex flexrow">
-        <label for="what">{{{localize 'swnr.dialog.failure.what'}}}</label>
-        <select name="what">
-          <option value="dest-all">{{localize 'swnr.dialog.failure.dest-all'}}</option>
-          <option value="break-all">{{localize 'swnr.dialog.failure.break-all'}}</option>
-          <option value="all-50">{{localize 'swnr.dialog.failure.all-50'}}</option>
-          <option value="break-1" selected>{{localize 'swnr.dialog.failure.break-1'}}</option>
-        </select>    
-      </div>
-  </div>
-</form>
