--- templates/dialogs/roll-spike.hbs
@@ -1,77 +1,0 @@
-<form >
-  <div class="flex flex-col">
-    <div class="grid grid-cols-4 gap-4">
-      <div class="flex flex-col">
-        <label for="pilotId">{{localize 'swnr.sheet.ship.station.bridge'}}:</label>
-        <select name="pilotId">
-          {{#select pilotId}}
-          {{#each crewArray }}
-          <option value="{{this.id}}">{{this.name}}</option>
-          {{/each}}
-          {{/select}}
-        </select>
-      </div>
-
-      <div class="flex flex-col">
-        <label for="skill">{{localize 'swnr.sheet.skill'}}: </label>
-        <select name="skill">
-          <option value="">{{ localize 'swnr.weapon.noSetSkill'}}</option>
-          {{#each pilot.itemTypes.skill }}
-          <option value="{{this.name}}" 
-            {{#if (eq this.name ../defaultSkill1)}} selected {{/if}} 
-            {{#if (eq this.name ../defaultSkill2)}} selected {{/if}}
-          >{{this.name}}</option>
-          {{/each}}
-        </select>
-      </div>
-      <div class="flex flex-col">
-        {{! dice }}
-        <label for="">{{localize 'swnr.chat.dicepool'}}: </label>
-        <select name="dicepool">
-          <option value="2d6">2D6</option>
-          <option value="3d6kh2">3D6</option>
-          <option value="4d6kh2">4D6</option>
-        </select>
-      </div>
-
-      <div class="flex flex-col">
-        {{! stat}}
-        <label for="stat">{{localize 'swnr.sheet.stats'}}: </label>
-        <select name="stat">
-          {{#select defaultStat}}
-          {{#each pilot.system.stats as |box name |}}
-          <option value="{{name}}">{{localize (concat "swnr.stat.long." name)}}</option>
-          {{/each}}
-          {{/select}}
-        </select>
-      </div>
-
-      <div class="flex flex-col">
-        <label for="modifier">{{localize 'swnr.chat.modifier'}}: </label>
-        <input name="modifier" type="number" step="1" value="0" />
-      </div>
-
-      <div class="flex flex-col">
-        <label for="travelDays">{{localize 'swnr.sheet.ship.travel'}}: </label>
-        <input name="travelDays" type="number" step="1" value="0" />
-      </div>
-
-      <div class="flex flex-col">
-        <label for="difficulty">{{localize 'swnr.chat.difficulty'}}: </label>
-        <input name="difficulty" type="number" step="1" value="{{baseDifficulty}}" />
-      </div>
-    </div>
-    <div>
-      Difficulty Modifiers:
-      <ul>
-        <li>+6 Uncharted</li>
-        <li>+2 Rutter more than 5 years </li>
-        <li>+1 Rutter is from 1-5 years old</li>
-        <li>-2 Rutter is less than 1 month old</li>
-        <li>+1 per 2 full hexes </li>
-        <li>+2 Trimming the course</li>
-        <li>+2 Rushed drill activation (one round vs 30 min)</li>
-      </ul>
-    </div>
-  </div>
-</form>
